% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNGCircos.R
\docType{data}
\name{arcExample}
\alias{arcExample}
\title{Arc plot example data}
\format{
A data frame with 7 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{color}{color}
  \item{des}{description}
  \item{link}{hyperlink}
  \item{html}{The external html language}
}
}
\usage{
arcExample
}
\description{
The data is in matrix with row names
}
\keyword{datasets}
