% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\docType{data}
\name{bubbleExample}
\alias{bubbleExample}
\title{Bubble plot example data}
\format{
A data frame with 8 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{name}{name for description}
  \item{value}{value}
  \item{color}{specified color for bubble}
  \item{layer}{layer number}
  \item{html}{The external html language}
}
}
\usage{
bubbleExample
}
\description{
The data is in matrix with column names
}
\keyword{datasets}
