% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\docType{data}
\name{chord.pExample}
\alias{chord.pExample}
\title{Example data of chord plot of circosJS}
\format{
A data frame in which each row represents the relationship from one genome position(source) to 
another one(target):
\describe{
  \item{source_chr}{chromosome name of source}
  \item{source_start}{start position of source}
  \item{source_end}{end position of source}
  \item{target_chr}{chromosome name of target}
  \item{target_start}{start position of target}
  \item{target_end}{end position of target}
}
}
\usage{
chord.pExample
}
\description{
The data is in matrix with column names
}
\keyword{datasets}
