% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\docType{data}
\name{chordExample}
\alias{chordExample}
\title{Example data of chord plot of NG-Circos}
\format{
A data frame in which each value represents the relationship from a column to a row:
\describe{
  \item{C.CK}{Genome 1, the name for each arc}
  \item{C.NPK}{Genome 2, the name for each arc}
  \item{GC.CK}{Genome 2, the name for each arc}
  \item{GC.NPK}{Genome 2, the name for each arc}
  \item{Alphaproteobacteria}{Genome 2, the name for each arc}
  \item{Betaproteobacteria}{Genome 2, the name for each arc}
  \item{Gammaproteobacteria}{Genome 2, the name for each arc}
  \item{Deltaproteobacteria}{Genome 8, the name for each arc}
  \item{Acidobacteria}{Genome 9, the name for each arc}
  \item{Actinobacteria}{Genome 10, the name for each arc}
  \item{Bacteroidetes}{Genome 11, the name for each arc}
  \item{Chloroflexi}{Genome 12, the name for each arc}
  \item{Firmicutes}{Genome 13, the name for each arc}
  \item{Gemmatimonadetes}{Genome 14, the name for each arc}
  \item{Planctomycetes}{Genome 15, the name for each arc}
  \item{Thaumarchaeota}{Genome 16, the name for each arc}
  \item{Verrucomicrobia}{Genome 17, the name for each arc}
  \item{Ascomycota}{Genome 18, the name for each arc}
  \item{Basidiomycota}{Genome 19, the name for each arc}
  \item{Zygomycota}{Genome 20, the name for each arc}
}
}
\usage{
chordExample
}
\description{
The data is in matrix with column names. The order and number is same as column, representing the same items
}
\keyword{datasets}
