% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\name{CircosChord.p}
\alias{CircosChord.p}
\title{CHORD module of circosJS}
\usage{
CircosChord.p(
  modulename,
  radius = 216,
  opacity = 0.67,
  color = "#B8B8B8",
  data,
  ...
)
}
\arguments{
\item{modulename}{The name of the new module}

\item{radius}{The radius for chord circle}

\item{opacity}{The opacity for chord}

\item{color}{The color for chord}

\item{data}{A list of chord value with relationship details, details could be found on chord.pExample}

\item{...}{Ignored}
}
\value{
The module tracks for chord modules of circosJS
}
\description{
Create a chord module using a data path. chord.p meaens chord plot based on path
}
\examples{

chord.pData<-chord.pExample
Circos()

}
