% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\name{CircosModuleList}
\alias{CircosModuleList}
\alias{+.CircosModuleList}
\alias{-.CircosModuleList}
\title{Create a list of modules}
\usage{
CircosModuleList()

\method{+}{CircosModuleList}(x, ...)

\method{-}{CircosModuleList}(x, ...)
}
\arguments{
\item{x}{The moduleList on which other modules should be added or removed.}

\item{...}{The modules to add (as moduleLists) or to remove (as module names).}
}
\value{
The list of all tracks of modules.
}
\description{
This allows the use of the '+' and '-' operator on these lists
}
