% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNGCircos.R
\docType{data}
\name{cnvExample}
\alias{cnvExample}
\title{Cnv module example data}
\format{
A data frame with 7 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{value}{value}
  \item{link}{hyperlink for cnv}
  \item{color}{color}
  \item{html}{The external html language}
}
}
\usage{
cnvExample
}
\description{
The data is in matrix with row names
}
\keyword{datasets}
