% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNGCircos.R
\docType{data}
\name{geneExample}
\alias{geneExample}
\title{Gene plot example data}
\format{
A data frame with 8 columns:
\describe{
  \item{chr}{chromosome}
  \item{strand}{strand, - or +}
  \item{start}{start position}
  \item{end}{end position}
  \item{type}{region type, gene or utr or cds}
  \item{name}{name for description}
  \item{link}{hyperlink for this region}
  \item{html}{The external html language}
}
}
\usage{
geneExample
}
\description{
The data is in matrix with row names
}
\keyword{datasets}
