% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCircos.R
\docType{data}
\name{heatmapExample}
\alias{heatmapExample}
\title{Heatmap plot example data}
\format{
A data frame with 7 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{name}{name for description}
  \item{value}{value}
  \item{layer}{layer number}
  \item{html}{The external html language}
}
}
\usage{
heatmapExample
}
\description{
The data is in matrix with row names
}
\keyword{datasets}
