% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\docType{data}
\name{hg19_ideogram}
\alias{hg19_ideogram}
\title{Ideogram for hg19}
\format{
A data frame with 4 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{color}{color}
}
}
\usage{
hg19_ideogram
}
\description{
The ideogram for human hg19 reference including the color for each region
}
\keyword{datasets}
