% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\docType{data}
\name{histogramExample}
\alias{histogramExample}
\title{Histogram plot example data}
\format{
A data frame with 7 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{name}{name for description}
  \item{link}{hyperlink}
  \item{value}{value}
  \item{html}{The external html language}
}
}
\usage{
histogramExample
}
\description{
The data is in matrix with column names
}
\keyword{datasets}
