% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCircos.R
\docType{data}
\name{lineExample}
\alias{lineExample}
\title{Line plot example data}
\format{
A data frame with 5 columns:
\describe{
  \item{chr}{chromosome}
  \item{pos}{position}
  \item{des}{description}
  \item{value}{value}
  \item{html}{The external html language}
}
}
\usage{
lineExample
}
\description{
The data is in matrix with row names
}
\keyword{datasets}
