% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCircos.R
\docType{data}
\name{linkExample}
\alias{linkExample}
\title{Link plot example data}
\format{
A data frame with 11 columns:
\describe{
  \item{g1chr}{first chromosome}
  \item{g1start}{first start position}
  \item{g1end}{first end position}
  \item{g2chr}{second chromosome}
  \item{g2start}{second start position}
  \item{g2end}{second end position}
  \item{g1name}{first name}
  \item{g2name}{second name}
  \item{fusion}{fusion name}
  \item{link}{hyperlink for link line}
  \item{html}{The external html language}
}
}
\usage{
linkExample
}
\description{
The data is in matrix with row names
}
\keyword{datasets}
