% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\docType{data}
\name{lollipopExample}
\alias{lollipopExample}
\title{Lollipop plot example data}
\format{
A data frame with 12 columns:
\describe{
  \item{protein}{protein name}
  \item{chr}{chromosome}
  \item{pos}{position}
  \item{strand}{strand, - or +}
  \item{CancerTypeNumber}{Cancer type number}
  \item{color}{color}
  \item{link}{hyperlink}
  \item{Consequence}{consequence}
  \item{AA_pos}{AA_pos}
  \item{AA_change}{AA_change}
  \item{type}{type for mutation,Hetero or Homo}
  \item{html}{The external html language}
}
}
\usage{
lollipopExample
}
\description{
The data is in matrix with column names
}
\keyword{datasets}
