% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCircos.R
\name{NGCircos}
\alias{NGCircos}
\title{NGCircos}
\usage{
NGCircos(
  tracklist = NGCircosTracklist(),
  genome = "hg19",
  genome2 = "hg19",
  genomeFillColor = "Spectral",
  chrPad = 0.02,
  width = NULL,
  height = NULL,
  innerRadius = 216,
  outerRadius = 240,
  svgClassName = "NGCircos",
  displayGenomeBorder = TRUE,
  genomeBorderColor = "#000",
  genomeBorderSize = 0.5,
  genomeTicksDisplay = FALSE,
  genomeTicksLen = 5,
  genomeTicksColor = "#000",
  genomeTicksTextSize = "0.6em",
  genomeTicksRealLength = TRUE,
  genomeTicksTextColor = "#000",
  genomeTicksScale = 3e+07,
  genomeTicksOffset = 0,
  genomeLabelDisplay = TRUE,
  genomeLabelTextSize = "10pt",
  genomeLabelTextColor = "#000",
  genomeLabelDx = 0,
  genomeLabelDy = 0,
  compareEvent = FALSE,
  compareEventGroupGapRate = 0.1,
  compareEventGroupDistance = 0,
  zoom = TRUE,
  TEXTModuleDragEvent = FALSE,
  CNVxlink = FALSE,
  CNVMouseEvent = TRUE,
  CNVMouseClickDisplay = FALSE,
  CNVMouseClickColor = "red",
  CNVMouseClickArcOpacity = 1,
  CNVMouseClickArcStrokeColor = "#F26223",
  CNVMouseClickArcStrokeWidth = 0,
  CNVMouseClickTextFromData = "fourth",
  CNVMouseClickTextOpacity = 1,
  CNVMouseClickTextColor = "red",
  CNVMouseClickTextSize = 8,
  CNVMouseClickTextPostionX = 0,
  CNVMouseClickTextPostionY = 0,
  CNVMouseClickTextDrag = TRUE,
  CNVMouseDownDisplay = FALSE,
  CNVMouseDownColor = "green",
  CNVMouseDownArcOpacity = 1,
  CNVMouseDownArcStrokeColor = "#F26223",
  CNVMouseDownArcStrokeWidth = 0,
  CNVMouseEnterDisplay = FALSE,
  CNVMouseEnterColor = "yellow",
  CNVMouseEnterArcOpacity = 1,
  CNVMouseEnterArcStrokeColor = "#F26223",
  CNVMouseEnterArcStrokeWidth = 0,
  CNVMouseLeaveDisplay = FALSE,
  CNVMouseLeaveColor = "pink",
  CNVMouseLeaveArcOpacity = 1,
  CNVMouseLeaveArcStrokeColor = "#F26223",
  CNVMouseLeaveArcStrokeWidth = 0,
  CNVMouseMoveDisplay = FALSE,
  CNVMouseMoveColor = "red",
  CNVMouseMoveArcOpacity = 1,
  CNVMouseMoveArcStrokeColor = "#F26223",
  CNVMouseMoveArcStrokeWidth = 0,
  CNVMouseOutDisplay = FALSE,
  CNVMouseOutAnimationTime = 500,
  CNVMouseOutColor = "red",
  CNVMouseOutArcOpacity = 1,
  CNVMouseOutArcStrokeColor = "red",
  CNVMouseOutArcStrokeWidth = 0,
  CNVMouseUpDisplay = FALSE,
  CNVMouseUpColor = "grey",
  CNVMouseUpArcOpacity = 1,
  CNVMouseUpArcStrokeColor = "#F26223",
  CNVMouseUpArcStrokeWidth = 0,
  CNVMouseOverDisplay = FALSE,
  CNVMouseOverColor = "red",
  CNVMouseOverArcOpacity = 1,
  CNVMouseOverArcStrokeColor = "#F26223",
  CNVMouseOverArcStrokeWidth = 3,
  CNVMouseOverTooltipsSetting = "style1",
  CNVMouseOverTooltipsHtml = " ",
  CNVMouseOverTooltipsPosition = "absolute",
  CNVMouseOverTooltipsBackgroundColor = "white",
  CNVMouseOverTooltipsBorderStyle = "solid",
  CNVMouseOverTooltipsBorderWidth = 0,
  CNVMouseOverTooltipsPadding = "3px",
  CNVMouseOverTooltipsBorderRadius = "3px",
  CNVMouseOverTooltipsOpacity = 0.8,
  HEATMAPMouseEvent = TRUE,
  HEATMAPMouseClickDisplay = FALSE,
  HEATMAPMouseClickColor = "green",
  HEATMAPMouseClickOpacity = 1,
  HEATMAPMouseClickStrokeColor = "none",
  HEATMAPMouseClickStrokeWidth = "none",
  HEATMAPMouseDownDisplay = FALSE,
  HEATMAPMouseDownColor = "green",
  HEATMAPMouseDownOpacity = 1,
  HEATMAPMouseDownStrokeColor = "none",
  HEATMAPMouseDownStrokeWidth = "none",
  HEATMAPMouseEnterDisplay = FALSE,
  HEATMAPMouseEnterColor = "green",
  HEATMAPMouseEnterOpacity = 1,
  HEATMAPMouseEnterStrokeColor = "none",
  HEATMAPMouseEnterStrokeWidth = "none",
  HEATMAPMouseLeaveDisplay = FALSE,
  HEATMAPMouseLeaveColor = "green",
  HEATMAPMouseLeaveOpacity = 1,
  HEATMAPMouseLeaveStrokeColor = "none",
  HEATMAPMouseLeaveStrokeWidth = "none",
  HEATMAPMouseMoveDisplay = FALSE,
  HEATMAPMouseMoveColor = "green",
  HEATMAPMouseMoveOpacity = 1,
  HEATMAPMouseMoveStrokeColor = "none",
  HEATMAPMouseMoveStrokeWidth = "none",
  HEATMAPMouseOutDisplay = FALSE,
  HEATMAPMouseOutAnimationTime = 500,
  HEATMAPMouseOutColor = "green",
  HEATMAPMouseOutOpacity = 1,
  HEATMAPMouseOutStrokeColor = "none",
  HEATMAPMouseOutStrokeWidth = "none",
  HEATMAPMouseUpDisplay = FALSE,
  HEATMAPMouseUpColor = "green",
  HEATMAPMouseUpOpacity = 1,
  HEATMAPMouseUpStrokeColor = "none",
  HEATMAPMouseUpStrokeWidth = "none",
  HEATMAPMouseOverDisplay = FALSE,
  HEATMAPMouseOverColor = "none",
  HEATMAPMouseOverOpacity = 1,
  HEATMAPMouseOverStrokeColor = "none",
  HEATMAPMouseOverStrokeWidth = "none",
  HEATMAPMouseOverTooltipsSetting = "style1",
  HEATMAPMouseOverTooltipsHtml = " ",
  HEATMAPMouseOverTooltipsPosition = "absolute",
  HEATMAPMouseOverTooltipsBackgroundColor = "white",
  HEATMAPMouseOverTooltipsBorderStyle = "solid",
  HEATMAPMouseOverTooltipsBorderWidth = 0,
  HEATMAPMouseOverTooltipsPadding = "3px",
  HEATMAPMouseOverTooltipsBorderRadius = "3px",
  HEATMAPMouseOverTooltipsOpacity = 0.8,
  BUBBLExlink = FALSE,
  BUBBLEMouseEvent = TRUE,
  BUBBLEMouseClickDisplay = FALSE,
  BUBBLEMouseClickColor = "green",
  BUBBLEMouseClickOpacity = 1,
  BUBBLEMouseClickStrokeColor = "none",
  BUBBLEMouseClickStrokeWidth = "none",
  BUBBLEMouseDownDisplay = FALSE,
  BUBBLEMouseDownColor = "green",
  BUBBLEMouseDownOpacity = 1,
  BUBBLEMouseDownStrokeColor = "none",
  BUBBLEMouseDownStrokeWidth = "none",
  BUBBLEMouseEnterDisplay = FALSE,
  BUBBLEMouseEnterColor = "green",
  BUBBLEMouseEnterOpacity = 1,
  BUBBLEMouseEnterStrokeColor = "none",
  BUBBLEMouseEnterStrokeWidth = "none",
  BUBBLEMouseLeaveDisplay = FALSE,
  BUBBLEMouseLeaveColor = "green",
  BUBBLEMouseLeaveOpacity = 1,
  BUBBLEMouseLeaveStrokeColor = "none",
  BUBBLEMouseLeaveStrokeWidth = "none",
  BUBBLEMouseMoveDisplay = FALSE,
  BUBBLEMouseMoveColor = "green",
  BUBBLEMouseMoveOpacity = 1,
  BUBBLEMouseMoveStrokeColor = "none",
  BUBBLEMouseMoveStrokeWidth = "none",
  BUBBLEMouseOutDisplay = FALSE,
  BUBBLEMouseOutAnimationTime = 500,
  BUBBLEMouseOutColor = "green",
  BUBBLEMouseOutOpacity = 1,
  BUBBLEMouseOutStrokeColor = "none",
  BUBBLEMouseOutStrokeWidth = "none",
  BUBBLEMouseUpDisplay = FALSE,
  BUBBLEMouseUpColor = "green",
  BUBBLEMouseUpOpacity = 1,
  BUBBLEMouseUpStrokeColor = "none",
  BUBBLEMouseUpStrokeWidth = "none",
  BUBBLEMouseOverDisplay = FALSE,
  BUBBLEMouseOverColor = "green",
  BUBBLEMouseOverOpacity = 1,
  BUBBLEMouseOverStrokeColor = "none",
  BUBBLEMouseOverStrokeWidth = "none",
  BUBBLEMouseOverTooltipsSetting = "style1",
  BUBBLEMouseOverTooltipsHtml = " ",
  BUBBLEMouseOverTooltipsPosition = "absolute",
  BUBBLEMouseOverTooltipsBackgroundColor = "white",
  BUBBLEMouseOverTooltipsBorderStyle = "solid",
  BUBBLEMouseOverTooltipsBorderWidth = 0,
  BUBBLEMouseOverTooltipsPadding = "3px",
  BUBBLEMouseOverTooltipsBorderRadius = "3px",
  BUBBLEMouseOverTooltipsOpacity = 0.8,
  SNPxlink = FALSE,
  SNPMouseEvent = TRUE,
  SNPMouseCombinationEvent = FALSE,
  SNPMouseCombinationImageDisplay = FALSE,
  SNPMouseCombinationImageTitle = "This is image",
  SNPMouseCombinationImageTitleSize = 5,
  SNPMouseCombinationImageTitleWeight = "bold",
  SNPMouseCombinationImageTitleColor = "black",
  SNPMouseCombinationImagePositionX = 0,
  SNPMouseCombinationImagePositionY = 0,
  SNPMouseCombinationImageHeight = 200,
  SNPMouseCombinationImageWidth = 300,
  SNPMouseCombinationGraphDisplay = FALSE,
  SNPMouseCombinationGraphTitle = "This is graph",
  SNPMouseCombinationGraphTitleSize = 5,
  SNPMouseCombinationGraphTitleWeight = "bold",
  SNPMouseCombinationGraphTitleColor = "black",
  SNPMouseCombinationGraphType = "histogram",
  SNPMouseCombinationGraphPositionX = 0,
  SNPMouseCombinationGraphPositionY = 0,
  SNPMouseCombinationGraphHeight = 200,
  SNPMouseCombinationGraphWidth = 300,
  SNPMouseCombinationGraphHistogramBarColor = "blue",
  SNPMouseCombinationGraphHistogramPadding = 30,
  SNPMouseCombinationGraphHistogramPositionCorrectX = 0,
  SNPMouseCombinationGraphPieAutoColor = TRUE,
  SNPMouseCombinationGraphPieColor = c("blue", "orange"),
  SNPMouseCombinationGraphPieSize = 50,
  SNPMouseCombinationGraphPieStroke = TRUE,
  SNPMouseCombinationGraphPieStrokeColor = "black",
  SNPMouseCombinationGraphPieStrokeWidth = 1,
  SNPMouseCombinationGraphPieOpacity = 1,
  SNPMouseCombinationGraphLineType = "linear",
  SNPMouseCombinationGraphLineColor = "black",
  SNPMouseCombinationGraphLineWidth = 1,
  SNPMouseCombinationGraphLinePoint = FALSE,
  SNPMouseCombinationGraphLinePointSize = 5,
  SNPMouseCombinationGraphLinePointAutoColor = TRUE,
  SNPMouseCombinationGraphLinePointColor = c("blue", "orange"),
  SNPMouseCombinationGraphLinePointStroke = TRUE,
  SNPMouseCombinationGraphLinePointStrokeColor = "black",
  SNPMouseCombinationGraphLinePointStrokeWidth = 1,
  SNPMouseCombinationGraphLinePointOpacity = 1,
  SNPMouseCombinationGraphLinePositionCorrectX = 0,
  SNPMouseCombinationTextDisplay = FALSE,
  SNPMouseCombinationTextColor = "red",
  SNPMouseCombinationTextSize = 3,
  SNPMouseCombinationTextWeight = "bold",
  SNPMouseCombinationTextPositionCorrectX = 0,
  SNPMouseCombinationTextPositionCorrectY = 0,
  SNPMouseClickDisplay = FALSE,
  SNPMouseClickColor = "red",
  SNPMouseClickCircleSize = 4,
  SNPMouseClickCircleOpacity = 1,
  SNPMouseClickCircleStrokeColor = "#F26223",
  SNPMouseClickCircleStrokeWidth = 0,
  SNPMouseClickTextFromData = "fourth",
  SNPMouseClickTextOpacity = 1,
  SNPMouseClickTextColor = "red",
  SNPMouseClickTextSize = 8,
  SNPMouseClickTextPostionX = 1,
  SNPMouseClickTextPostionY = 10,
  SNPMouseClickTextDrag = TRUE,
  SNPMouseDownDisplay = FALSE,
  SNPMouseDownColor = "green",
  SNPMouseDownCircleSize = 4,
  SNPMouseDownCircleOpacity = 1,
  SNPMouseDownCircleStrokeColor = "#F26223",
  SNPMouseDownCircleStrokeWidth = 0,
  SNPMouseEnterDisplay = FALSE,
  SNPMouseEnterColor = "yellow",
  SNPMouseEnterCircleSize = 4,
  SNPMouseEnterCircleOpacity = 1,
  SNPMouseEnterCircleStrokeColor = "#F26223",
  SNPMouseEnterCircleStrokeWidth = 0,
  SNPMouseLeaveDisplay = FALSE,
  SNPMouseLeaveColor = "pink",
  SNPMouseLeaveCircleSize = 4,
  SNPMouseLeaveCircleOpacity = 1,
  SNPMouseLeaveCircleStrokeColor = "#F26223",
  SNPMouseLeaveCircleStrokeWidth = 0,
  SNPMouseMoveDisplay = FALSE,
  SNPMouseMoveColor = "red",
  SNPMouseMoveCircleSize = 2,
  SNPMouseMoveCircleOpacity = 1,
  SNPMouseMoveCircleStrokeColor = "#F26223",
  SNPMouseMoveCircleStrokeWidth = 0,
  SNPMouseOutDisplay = FALSE,
  SNPMouseOutAnimationTime = 500,
  SNPMouseOutColor = "red",
  SNPMouseOutCircleSize = 2,
  SNPMouseOutCircleOpacity = 1,
  SNPMouseOutCircleStrokeColor = "red",
  SNPMouseOutCircleStrokeWidth = 0,
  SNPMouseUpDisplay = FALSE,
  SNPMouseUpColor = "grey",
  SNPMouseUpCircleSize = 2,
  SNPMouseUpCircleOpacity = 1,
  SNPMouseUpCircleStrokeColor = "#F26223",
  SNPMouseUpCircleStrokeWidth = 0,
  SNPMouseOverDisplay = FALSE,
  SNPMouseOverColor = "red",
  SNPMouseOverCircleSize = 2,
  SNPMouseOverCircleOpacity = 1,
  SNPMouseOverCircleStrokeColor = "#F26223",
  SNPMouseOverCircleStrokeWidth = 3,
  SNPMouseOverTooltipsSetting = "style1",
  SNPMouseOverTooltipsHtml = " ",
  SNPMouseOverTooltipsPosition = "absolute",
  SNPMouseOverTooltipsBackgroundColor = "white",
  SNPMouseOverTooltipsBorderStyle = "solid",
  SNPMouseOverTooltipsBorderWidth = 0,
  SNPMouseOverTooltipsPadding = "3px",
  SNPMouseOverTooltipsBorderRadius = "3px",
  SNPMouseOverTooltipsOpacity = 0.8,
  LINKxlink = FALSE,
  LINKMouseEvent = TRUE,
  LINKMouseClickDisplay = FALSE,
  LINKMouseClickOpacity = 1,
  LINKMouseClickStrokeColor = "green",
  LINKMouseClickStrokeWidth = 4,
  LINKMouseDownDisplay = FALSE,
  LINKMouseDownOpacity = 1,
  LINKMouseDownStrokeColor = "none",
  LINKMouseDownStrokeWidth = "none",
  LINKMouseEnterDisplay = FALSE,
  LINKMouseEnterOpacity = 1,
  LINKMouseEnterStrokeColor = "none",
  LINKMouseEnterStrokeWidth = "none",
  LINKMouseLeaveDisplay = FALSE,
  LINKMouseLeaveOpacity = 1,
  LINKMouseLeaveStrokeColor = "none",
  LINKMouseLeaveStrokeWidth = "none",
  LINKMouseMoveDisplay = FALSE,
  LINKMouseMoveOpacity = 1,
  LINKMouseMoveStrokeColor = "none",
  LINKMouseMoveStrokeWidth = "none",
  LINKMouseOutDisplay = FALSE,
  LINKMouseOutAnimationTime = 500,
  LINKMouseOutOpacity = 1,
  LINKMouseOutStrokeColor = "none",
  LINKMouseOutStrokeWidth = "none",
  LINKMouseUpDisplay = FALSE,
  LINKMouseUpOpacity = 1,
  LINKMouseUpStrokeColor = "none",
  LINKMouseUpStrokeWidth = "none",
  LINKMouseOverDisplay = FALSE,
  LINKMouseOverOpacity = 1,
  LINKMouseOverStrokeColor = "none",
  LINKMouseOverStrokeWidth = "none",
  LINKMouseOverTooltipsSetting = "style1",
  LINKMouseOverTooltipsHtml = " ",
  LINKMouseOverTooltipsPosition = "absolute",
  LINKMouseOverTooltipsBackgroundColor = "white",
  LINKMouseOverTooltipsBorderStyle = "solid",
  LINKMouseOverTooltipsBorderWidth = 0,
  LINKMouseOverTooltipsPadding = "3px",
  LINKMouseOverTooltipsBorderRadius = "3px",
  LINKMouseOverTooltipsOpacity = 1,
  LINKLabelDragEvent = FALSE,
  CHORDMouseEvent = TRUE,
  CHORDMouseFillColorExcluded = "#FFFFFF",
  CHORDMouseClickDisplay = FALSE,
  CHORDMouseClickOpacity = 1,
  CHORDMouseClickStrokeColor = "none",
  CHORDMouseClickStrokeWidth = "none",
  CHORDMouseDownDisplay = FALSE,
  CHORDMouseDownOpacity = 1,
  CHORDMouseDownStrokeColor = "none",
  CHORDMouseDownStrokeWidth = "none",
  CHORDMouseEnterDisplay = FALSE,
  CHORDMouseEnterOpacity = 1,
  CHORDMouseEnterStrokeColor = "none",
  CHORDMouseEnterStrokeWidth = "none",
  CHORDMouseLeaveDisplay = FALSE,
  CHORDMouseLeaveOpacity = 1,
  CHORDMouseLeaveStrokeColor = "none",
  CHORDMouseLeaveStrokeWidth = "none",
  CHORDMouseMoveDisplay = FALSE,
  CHORDMouseMoveOpacity = 1,
  CHORDMouseMoveStrokeColor = "none",
  CHORDMouseMoveStrokeWidth = "none",
  CHORDMouseOutDisplay = FALSE,
  CHORDMouseOutAnimationTime = 500,
  CHORDMouseOutOpacity = 1,
  CHORDMouseOutStrokeColor = "none",
  CHORDMouseOutStrokeWidth = "none",
  CHORDMouseUpDisplay = FALSE,
  CHORDMouseUpOpacity = 1,
  CHORDMouseUpStrokeColor = "none",
  CHORDMouseUpStrokeWidth = "none",
  CHORDMouseOverDisplay = FALSE,
  CHORDMouseOverOpacity = 1,
  CHORDMouseOverStrokeColor = "none",
  CHORDMouseOverStrokeWidth = "none",
  HISTOGRAMxlink = FALSE,
  HISTOGRAMMouseEvent = TRUE,
  HISTOGRAMMouseClickDisplay = FALSE,
  HISTOGRAMMouseClickColor = "red",
  HISTOGRAMMouseClickOpacity = 1,
  HISTOGRAMMouseClickStrokeColor = "none",
  HISTOGRAMMouseClickStrokeWidth = "none",
  HISTOGRAMMouseDownDisplay = FALSE,
  HISTOGRAMMouseDownColor = "red",
  HISTOGRAMMouseDownOpacity = 1,
  HISTOGRAMMouseDownStrokeColor = "none",
  HISTOGRAMMouseDownStrokeWidth = "none",
  HISTOGRAMMouseEnterDisplay = FALSE,
  HISTOGRAMMouseEnterColor = "red",
  HISTOGRAMMouseEnterOpacity = 1,
  HISTOGRAMMouseEnterStrokeColor = "none",
  HISTOGRAMMouseEnterStrokeWidth = "none",
  HISTOGRAMMouseLeaveDisplay = FALSE,
  HISTOGRAMMouseLeaveColor = "red",
  HISTOGRAMMouseLeaveOpacity = 1,
  HISTOGRAMMouseLeaveStrokeColor = "none",
  HISTOGRAMMouseLeaveStrokeWidth = "none",
  HISTOGRAMMouseMoveDisplay = FALSE,
  HISTOGRAMMouseMoveColor = "red",
  HISTOGRAMMouseMoveOpacity = 1,
  HISTOGRAMMouseMoveStrokeColor = "none",
  HISTOGRAMMouseMoveStrokeWidth = "none",
  HISTOGRAMMouseOutDisplay = FALSE,
  HISTOGRAMMouseOutAnimationTime = 500,
  HISTOGRAMMouseOutColor = "red",
  HISTOGRAMMouseOutOpacity = 1,
  HISTOGRAMMouseOutStrokeColor = "none",
  HISTOGRAMMouseOutStrokeWidth = "none",
  HISTOGRAMMouseUpDisplay = FALSE,
  HISTOGRAMMouseUpColor = "red",
  HISTOGRAMMouseUpOpacity = 1,
  HISTOGRAMMouseUpStrokeColor = "none",
  HISTOGRAMMouseUpStrokeWidth = "none",
  HISTOGRAMMouseOverDisplay = FALSE,
  HISTOGRAMMouseOverColor = "red",
  HISTOGRAMMouseOverOpacity = 1,
  HISTOGRAMMouseOverStrokeColor = "none",
  HISTOGRAMMouseOverStrokeWidth = "none",
  HISTOGRAMMouseOverTooltipsSetting = "style1",
  HISTOGRAMMouseOverTooltipsHtml = " ",
  HISTOGRAMMouseOverTooltipsPosition = "absolute",
  HISTOGRAMMouseOverTooltipsBackgroundColor = "white",
  HISTOGRAMMouseOverTooltipsBorderStyle = "solid",
  HISTOGRAMMouseOverTooltipsBorderWidth = 0,
  HISTOGRAMMouseOverTooltipsPadding = "3px",
  HISTOGRAMMouseOverTooltipsBorderRadius = "3px",
  HISTOGRAMMouseOverTooltipsOpacity = 1,
  LINEMouseEvent = TRUE,
  LINEMouseClickDisplay = FALSE,
  LINEMouseClickLineOpacity = 1,
  LINEMouseClickLineStrokeColor = "none",
  LINEMouseClickLineStrokeWidth = "none",
  LINEMouseDownDisplay = FALSE,
  LINEMouseDownLineOpacity = 1,
  LINEMouseDownLineStrokeColor = "none",
  LINEMouseDownLineStrokeWidth = "none",
  LINEMouseEnterDisplay = FALSE,
  LINEMouseEnterLineOpacity = 1,
  LINEMouseEnterLineStrokeColor = "none",
  LINEMouseEnterLineStrokeWidth = "none",
  LINEMouseLeaveDisplay = FALSE,
  LINEMouseLeaveLineOpacity = 1,
  LINEMouseLeaveLineStrokeColor = "none",
  LINEMouseLeaveLineStrokeWidth = "none",
  LINEMouseMoveDisplay = FALSE,
  LINEMouseMoveLineOpacity = 1,
  LINEMouseMoveLineStrokeColor = "none",
  LINEMouseMoveLineStrokeWidth = "none",
  LINEMouseOutDisplay = FALSE,
  LINEMouseOutAnimationTime = 500,
  LINEMouseOutLineOpacity = 1,
  LINEMouseOutLineStrokeColor = "none",
  LINEMouseOutLineStrokeWidth = "none",
  LINEMouseUpDisplay = FALSE,
  LINEMouseUpLineOpacity = 1,
  LINEMouseUpLineStrokeColor = "none",
  LINEMouseUpLineStrokeWidth = "none",
  LINEMouseOverDisplay = FALSE,
  LINEMouseOverLineOpacity = 1,
  LINEMouseOverLineStrokeColor = "none",
  LINEMouseOverLineStrokeWidth = "none",
  LINEMouseOverTooltipsSetting = "style1",
  LINEMouseOverTooltipsHtml = " ",
  LINEMouseOverTooltipsPosition = "absolute",
  LINEMouseOverTooltipsBackgroundColor = "white",
  LINEMouseOverTooltipsBorderStyle = "solid",
  LINEMouseOverTooltipsBorderWidth = 0,
  LINEMouseOverTooltipsPadding = "3px",
  LINEMouseOverTooltipsBorderRadius = "3px",
  LINEMouseOverTooltipsOpacity = 1,
  WIGMouseEvent = TRUE,
  WIGMouseClickDisplay = FALSE,
  WIGMouseClickLineOpacity = 1,
  WIGMouseClickLineStrokeColor = "none",
  WIGMouseClickLineStrokeWidth = "none",
  WIGMouseClickFillColor = "none",
  WIGMouseDownDisplay = FALSE,
  WIGMouseDownLineOpacity = 1,
  WIGMouseDownLineStrokeColor = "none",
  WIGMouseDownLineStrokeWidth = "none",
  WIGMouseDownFillColor = "none",
  WIGMouseEnterDisplay = FALSE,
  WIGMouseEnterLineOpacity = 1,
  WIGMouseEnterLineStrokeColor = "none",
  WIGMouseEnterLineStrokeWidth = "none",
  WIGMouseEnterFillColor = "none",
  WIGMouseLeaveDisplay = FALSE,
  WIGMouseLeaveLineOpacity = 1,
  WIGMouseLeaveLineStrokeColor = "none",
  WIGMouseLeaveLineStrokeWidth = "none",
  WIGMouseLeaveFillColor = "none",
  WIGMouseMoveDisplay = FALSE,
  WIGMouseMoveLineOpacity = 1,
  WIGMouseMoveLineStrokeColor = "none",
  WIGMouseMoveLineStrokeWidth = "none",
  WIGMouseMoveFillColor = "none",
  WIGMouseOutDisplay = FALSE,
  WIGMouseOutAnimationTime = 500,
  WIGMouseOutLineOpacity = 1,
  WIGMouseOutLineStrokeColor = "none",
  WIGMouseOutLineStrokeWidth = "none",
  WIGMouseOutFillColor = "none",
  WIGMouseUpDisplay = FALSE,
  WIGMouseUpLineOpacity = 1,
  WIGMouseUpLineStrokeColor = "none",
  WIGMouseUpLineStrokeWidth = "none",
  WIGMouseUpFillColor = "none",
  WIGMouseOverDisplay = FALSE,
  WIGMouseOverLineOpacity = 1,
  WIGMouseOverLineStrokeColor = "none",
  WIGMouseOverLineStrokeWidth = "none",
  WIGMouseOverFillColor = "none",
  WIGMouseOverTooltipsSetting = "style1",
  WIGMouseOverTooltipsHtml = " ",
  WIGMouseOverTooltipsPosition = "absolute",
  WIGMouseOverTooltipsBackgroundColor = "white",
  WIGMouseOverTooltipsBorderStyle = "solid",
  WIGMouseOverTooltipsBorderWidth = 0,
  WIGMouseOverTooltipsPadding = "3px",
  WIGMouseOverTooltipsBorderRadius = "3px",
  WIGMouseOverTooltipsOpacity = 1,
  SCATTERxlink = FALSE,
  SCATTERMouseEvent = TRUE,
  SCATTERMouseClickDisplay = FALSE,
  SCATTERMouseClickColor = "red",
  SCATTERMouseClickCircleSize = 2,
  SCATTERMouseClickCircleOpacity = 1,
  SCATTERMouseClickCircleStrokeColor = "none",
  SCATTERMouseClickCircleStrokeWidth = "none",
  SCATTERMouseClickTextFromData = "fourth",
  SCATTERMouseClickTextOpacity = 1,
  SCATTERMouseClickTextColor = "red",
  SCATTERMouseClickTextSize = 8,
  SCATTERMouseClickTextPostionX = 1,
  SCATTERMouseClickTextPostionY = 10,
  SCATTERMouseClickTextDrag = TRUE,
  SCATTERMouseDownDisplay = FALSE,
  SCATTERMouseDownColor = "red",
  SCATTERMouseDownCircleSize = 2,
  SCATTERMouseDownCircleOpacity = 1,
  SCATTERMouseDownCircleStrokeColor = "none",
  SCATTERMouseDownCircleStrokeWidth = "none",
  SCATTERMouseEnterDisplay = FALSE,
  SCATTERMouseEnterColor = "red",
  SCATTERMouseEnterCircleSize = 2,
  SCATTERMouseEnterCircleOpacity = 1,
  SCATTERMouseEnterCircleStrokeColor = "none",
  SCATTERMouseEnterCircleStrokeWidth = "none",
  SCATTERMouseLeaveDisplay = FALSE,
  SCATTERMouseLeaveColor = "red",
  SCATTERMouseLeaveCircleSize = 2,
  SCATTERMouseLeaveCircleOpacity = 1,
  SCATTERMouseLeaveCircleStrokeColor = "none",
  SCATTERMouseLeaveCircleStrokeWidth = "none",
  SCATTERMouseMoveDisplay = FALSE,
  SCATTERMouseMoveColor = "red",
  SCATTERMouseMoveCircleSize = 2,
  SCATTERMouseMoveCircleOpacity = 1,
  SCATTERMouseMoveCircleStrokeColor = "none",
  SCATTERMouseMoveCircleStrokeWidth = "none",
  SCATTERMouseOutDisplay = FALSE,
  SCATTERMouseOutAnimationTime = 500,
  SCATTERMouseOutColor = "red",
  SCATTERMouseOutCircleSize = 2,
  SCATTERMouseOutCircleOpacity = 1,
  SCATTERMouseOutCircleStrokeColor = "none",
  SCATTERMouseOutCircleStrokeWidth = "none",
  SCATTERMouseUpDisplay = FALSE,
  SCATTERMouseUpColor = "red",
  SCATTERMouseUpCircleSize = 2,
  SCATTERMouseUpCircleOpacity = 1,
  SCATTERMouseUpCircleStrokeColor = "none",
  SCATTERMouseUpCircleStrokeWidth = "none",
  SCATTERMouseOverDisplay = FALSE,
  SCATTERMouseOverColor = "red",
  SCATTERMouseOverCircleSize = 2,
  SCATTERMouseOverCircleOpacity = 1,
  SCATTERMouseOverCircleStrokeColor = "none",
  SCATTERMouseOverCircleStrokeWidth = "none",
  SCATTERMouseOverTooltipsSetting = "style1",
  SCATTERMouseOverTooltipsHtml = " ",
  SCATTERMouseOverTooltipsPosition = "absolute",
  SCATTERMouseOverTooltipsBackgroundColor = "white",
  SCATTERMouseOverTooltipsBorderStyle = "solid",
  SCATTERMouseOverTooltipsBorderWidth = 0,
  SCATTERMouseOverTooltipsPadding = "3px",
  SCATTERMouseOverTooltipsBorderRadius = "3px",
  SCATTERMouseOverTooltipsOpacity = 1,
  ARCxlink = FALSE,
  ARCMouseEvent = TRUE,
  ARCMouseClickDisplay = FALSE,
  ARCMouseClickColor = "red",
  ARCMouseClickArcOpacity = 1,
  ARCMouseClickArcStrokeColor = "none",
  ARCMouseClickArcStrokeWidth = "none",
  ARCMouseClickTextFromData = "fourth",
  ARCMouseClickTextOpacity = 1,
  ARCMouseClickTextColor = "red",
  ARCMouseClickTextSize = 8,
  ARCMouseClickTextPostionX = 1,
  ARCMouseClickTextPostionY = 10,
  ARCMouseClickTextDrag = TRUE,
  ARCMouseDownDisplay = FALSE,
  ARCMouseDownColor = "red",
  ARCMouseDownArcOpacity = 1,
  ARCMouseDownArcStrokeColor = "none",
  ARCMouseDownArcStrokeWidth = "none",
  ARCMouseEnterDisplay = FALSE,
  ARCMouseEnterColor = "red",
  ARCMouseEnterArcOpacity = 1,
  ARCMouseEnterArcStrokeColor = "none",
  ARCMouseEnterArcStrokeWidth = "none",
  ARCMouseLeaveDisplay = FALSE,
  ARCMouseLeaveColor = "red",
  ARCMouseLeaveArcOpacity = 1,
  ARCMouseLeaveArcStrokeColor = "none",
  ARCMouseLeaveArcStrokeWidth = "none",
  ARCMouseMoveDisplay = FALSE,
  ARCMouseMoveColor = "red",
  ARCMouseMoveArcOpacity = 1,
  ARCMouseMoveArcStrokeColor = "none",
  ARCMouseMoveArcStrokeWidth = "none",
  ARCMouseOutDisplay = FALSE,
  ARCMouseOutAnimationTime = 500,
  ARCMouseOutColor = "red",
  ARCMouseOutArcOpacity = 1,
  ARCMouseOutArcStrokeColor = "none",
  ARCMouseOutArcStrokeWidth = "none",
  ARCMouseUpDisplay = FALSE,
  ARCMouseUpColor = "red",
  ARCMouseUpArcOpacity = 1,
  ARCMouseUpArcStrokeColor = "none",
  ARCMouseUpArcStrokeWidth = "none",
  ARCMouseOverDisplay = FALSE,
  ARCMouseOverColor = "red",
  ARCMouseOverArcOpacity = 1,
  ARCMouseOverArcStrokeColor = "none",
  ARCMouseOverArcStrokeWidth = "none",
  ARCMouseOverTooltipsSetting = "style1",
  ARCMouseOverTooltipsHtml = " ",
  ARCMouseOverTooltipsPosition = "absolute",
  ARCMouseOverTooltipsBackgroundColor = "white",
  ARCMouseOverTooltipsBorderStyle = "solid",
  ARCMouseOverTooltipsBorderWidth = 0,
  ARCMouseOverTooltipsPadding = "3px",
  ARCMouseOverTooltipsBorderRadius = "3px",
  ARCMouseOverTooltipsOpacity = 1,
  GENExlink = FALSE,
  GENEMouseEvent = TRUE,
  GENEMouseClickDisplay = FALSE,
  GENEMouseClickColor = "red",
  GENEMouseClickArcOpacity = 1,
  GENEMouseClickArcStrokeColor = "none",
  GENEMouseClickArcStrokeWidth = "none",
  GENEMouseClickTextFromData = "fourth",
  GENEMouseClickTextOpacity = 1,
  GENEMouseClickTextColor = "red",
  GENEMouseClickTextSize = 8,
  GENEMouseClickTextPostionX = 1,
  GENEMouseClickTextPostionY = 10,
  GENEMouseClickTextDrag = TRUE,
  GENEMouseDownDisplay = FALSE,
  GENEMouseDownColor = "red",
  GENEMouseDownArcOpacity = 1,
  GENEMouseDownArcStrokeColor = "none",
  GENEMouseDownArcStrokeWidth = "none",
  GENEMouseEnterDisplay = FALSE,
  GENEMouseEnterColor = "red",
  GENEMouseEnterArcOpacity = 1,
  GENEMouseEnterArcStrokeColor = "none",
  GENEMouseEnterArcStrokeWidth = "none",
  GENEMouseLeaveDisplay = FALSE,
  GENEMouseLeaveColor = "red",
  GENEMouseLeaveArcOpacity = 1,
  GENEMouseLeaveArcStrokeColor = "none",
  GENEMouseLeaveArcStrokeWidth = "none",
  GENEMouseMoveDisplay = FALSE,
  GENEMouseMoveColor = "red",
  GENEMouseMoveArcOpacity = 1,
  GENEMouseMoveArcStrokeColor = "none",
  GENEMouseMoveArcStrokeWidth = "none",
  GENEMouseOutDisplay = FALSE,
  GENEMouseOutAnimationTime = 500,
  GENEMouseOutColor = "red",
  GENEMouseOutArcOpacity = 1,
  GENEMouseOutArcStrokeColor = "none",
  GENEMouseOutArcStrokeWidth = "none",
  GENEMouseUpDisplay = FALSE,
  GENEMouseUpColor = "red",
  GENEMouseUpArcOpacity = 1,
  GENEMouseUpArcStrokeColor = "none",
  GENEMouseUpArcStrokeWidth = "none",
  GENEMouseOverDisplay = FALSE,
  GENEMouseOverColor = "red",
  GENEMouseOverArcOpacity = 1,
  GENEMouseOverArcStrokeColor = "none",
  GENEMouseOverArcStrokeWidth = "none",
  GENEMouseOverTooltipsSetting = "style1",
  GENEMouseOverTooltipsHtml = " ",
  GENEMouseOverTooltipsPosition = "absolute",
  GENEMouseOverTooltipsBackgroundColor = "white",
  GENEMouseOverTooltipsBorderStyle = "solid",
  GENEMouseOverTooltipsBorderWidth = 0,
  GENEMouseOverTooltipsPadding = "3px",
  GENEMouseOverTooltipsBorderRadius = "3px",
  GENEMouseOverTooltipsOpacity = 1,
  LOLLIPOPxlink = FALSE,
  LOLLIPOPMouseEvent = TRUE,
  LOLLIPOPMouseClickDisplay = FALSE,
  LOLLIPOPMouseClickColor = "red",
  LOLLIPOPMouseClickCircleSize = 2,
  LOLLIPOPMouseClickCircleOpacity = 1,
  LOLLIPOPMouseClickCircleStrokeColor = "none",
  LOLLIPOPMouseClickCircleStrokeWidth = "none",
  LOLLIPOPMouseClickTextFromData = "fourth",
  LOLLIPOPMouseClickTextOpacity = 1,
  LOLLIPOPMouseClickTextColor = "red",
  LOLLIPOPMouseClickTextSize = 8,
  LOLLIPOPMouseClickTextPostionX = 1,
  LOLLIPOPMouseClickTextPostionY = 10,
  LOLLIPOPMouseClickTextDrag = TRUE,
  LOLLIPOPMouseDownDisplay = FALSE,
  LOLLIPOPMouseDownColor = "red",
  LOLLIPOPMouseDownCircleSize = 2,
  LOLLIPOPMouseDownCircleOpacity = 1,
  LOLLIPOPMouseDownCircleStrokeColor = "none",
  LOLLIPOPMouseDownCircleStrokeWidth = "none",
  LOLLIPOPMouseEnterDisplay = FALSE,
  LOLLIPOPMouseEnterColor = "red",
  LOLLIPOPMouseEnterCircleSize = 2,
  LOLLIPOPMouseEnterCircleOpacity = 1,
  LOLLIPOPMouseEnterCircleStrokeColor = "none",
  LOLLIPOPMouseEnterCircleStrokeWidth = "none",
  LOLLIPOPMouseLeaveDisplay = FALSE,
  LOLLIPOPMouseLeaveColor = "red",
  LOLLIPOPMouseLeaveCircleSize = 2,
  LOLLIPOPMouseLeaveCircleOpacity = 1,
  LOLLIPOPMouseLeaveCircleStrokeColor = "none",
  LOLLIPOPMouseLeaveCircleStrokeWidth = "none",
  LOLLIPOPMouseMoveDisplay = FALSE,
  LOLLIPOPMouseMoveColor = "red",
  LOLLIPOPMouseMoveCircleSize = 2,
  LOLLIPOPMouseMoveCircleOpacity = 1,
  LOLLIPOPMouseMoveCircleStrokeColor = "none",
  LOLLIPOPMouseMoveCircleStrokeWidth = "none",
  LOLLIPOPMouseOutDisplay = FALSE,
  LOLLIPOPMouseOutAnimationTime = 500,
  LOLLIPOPMouseOutColor = "red",
  LOLLIPOPMouseOutCircleSize = 2,
  LOLLIPOPMouseOutCircleOpacity = 1,
  LOLLIPOPMouseOutCircleStrokeColor = "none",
  LOLLIPOPMouseOutCircleStrokeWidth = "none",
  LOLLIPOPMouseUpDisplay = FALSE,
  LOLLIPOPMouseUpColor = "red",
  LOLLIPOPMouseUpCircleSize = 2,
  LOLLIPOPMouseUpCircleOpacity = 1,
  LOLLIPOPMouseUpCircleStrokeColor = "none",
  LOLLIPOPMouseUpCircleStrokeWidth = "none",
  LOLLIPOPMouseOverDisplay = FALSE,
  LOLLIPOPMouseOverColor = "red",
  LOLLIPOPMouseOverCircleSize = 2,
  LOLLIPOPMouseOverCircleOpacity = 1,
  LOLLIPOPMouseOverCircleStrokeColor = "none",
  LOLLIPOPMouseOverCircleStrokeWidth = "none",
  LOLLIPOPMouseOverTooltipsSetting = "style1",
  LOLLIPOPMouseOverTooltipsHtml = " ",
  LOLLIPOPMouseOverTooltipsPosition = "absolute",
  LOLLIPOPMouseOverTooltipsBackgroundColor = "white",
  LOLLIPOPMouseOverTooltipsBorderStyle = "solid",
  LOLLIPOPMouseOverTooltipsBorderWidth = 0,
  LOLLIPOPMouseOverTooltipsPadding = "3px",
  LOLLIPOPMouseOverTooltipsBorderRadius = "3px",
  LOLLIPOPMouseOverTooltipsOpacity = 1,
  elementId = NULL,
  ...
)
}
\arguments{
\item{tracklist}{A list of module to display.}

\item{genome}{A list of chromosome lengths to be used as reference for the visualization or 'hg19' to use
the chromosomes 1 to 22 and the sexual chromosomes according to the hg19 reference.}

\item{genome2}{Second genome when compare module is applied}

\item{genomeFillColor}{The color to display in each chromosome. Can be a RColorBrewer palette name used to
generate one color per chromosome, or a character object or vector of character objects stating RGB values in hexadecimal
format or base R colors. If the vector is shorter than the reference genome, values will be repeated.}

\item{chrPad}{Distance between chromosomes.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{innerRadius}{Default 216, Inner radius of chromosome}

\item{outerRadius}{Default 240, Outer radius of chromosome}

\item{svgClassName}{The svg class name}

\item{displayGenomeBorder, genomeBorderColor, genomeBorderSize}{Should the reference genome have borders?
If yes specify the color, in RGB hexadecimal format, and the thickness.}

\item{genomeTicksDisplay, genomeTicksLen, genomeTicksColor, genomeTicksTextSize, genomeTicksTextColor, genomeTicksScale, genomeTicksRealLength, genomeTicksOffset}{Should the refence genome have ticks, of which length, color (in hexadecimal RGB format), with labels in which font
size and color, and spaced by how many bases? whether show the real length of genome? The offset from real length}

\item{genomeLabelDisplay, genomeLabelTextSize, genomeLabelTextColor, genomeLabelDx, genomeLabelDy}{Should the reference genome have labels on each chromosome, in which font size and color?}

\item{compareEvent}{Default False, open/not COMPARE module}

\item{compareEventGroupGapRate}{Default 0.1, control the two-side gap rate on each group of genome}

\item{compareEventGroupDistance}{Default 0, distance between two groups of genome}

\item{zoom}{Is zooming and moving in the visualization allowed?}

\item{TEXTModuleDragEvent}{Are text annotations draggable?}

\item{CNVxlink}{Default False, add/not xlink for CNV module}

\item{CNVMouseEvent}{Default True, open/not open mouse event of CNV module}

\item{CNVMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a CNV point.}

\item{CNVMouseClickColor}{Color when mouse clicking}

\item{CNVMouseClickArcOpacity}{Arc opacity when mouse clicking the element}

\item{CNVMouseClickArcStrokeColor}{Arc stroke color when mouse clicking the element}

\item{CNVMouseClickArcStrokeWidth}{Arc stroke width when mouse clicking the element}

\item{CNVMouseClickTextFromData}{Text column when mouse clicking the element}

\item{CNVMouseClickTextOpacity}{Text opacity when mouse clicking the element}

\item{CNVMouseClickTextColor}{Text color when mouse clicking the element}

\item{CNVMouseClickTextSize}{Text size when mouse clicking the element}

\item{CNVMouseClickTextPostionX, CNVMouseClickTextPostionY}{Text coordinates when mouse clicking the element}

\item{CNVMouseClickTextDrag}{Whether text is draggable when mouse clicking the element}

\item{CNVMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a CNV point.}

\item{CNVMouseDownColor}{Color when mouse moving down the element}

\item{CNVMouseDownArcOpacity}{Arc opacity when mouse moving down the element}

\item{CNVMouseDownArcStrokeColor}{Arc stroke color when mouse moving down the element}

\item{CNVMouseDownArcStrokeWidth}{Arc stroke width when mouse moving down the element}

\item{CNVMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a CNV point.}

\item{CNVMouseEnterColor}{Color when mouse entering the element}

\item{CNVMouseEnterArcOpacity}{Arc opacity when mouse entering the element}

\item{CNVMouseEnterArcStrokeColor}{Arc stroke color when mouse entering the element}

\item{CNVMouseEnterArcStrokeWidth}{Arc stroke width when mouse entering the element}

\item{CNVMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a CNV point.}

\item{CNVMouseLeaveColor}{Color when mouse leaving the element}

\item{CNVMouseLeaveArcOpacity}{Arc opacity when mouse leaving the element}

\item{CNVMouseLeaveArcStrokeColor}{Arc stroke color when mouse leaving the element}

\item{CNVMouseLeaveArcStrokeWidth}{Arc stroke width when mouse leaving the element}

\item{CNVMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a CNV point.}

\item{CNVMouseMoveColor}{Color when mouse moving in the element}

\item{CNVMouseMoveArcOpacity}{Arc opacity when mouse moving in the element}

\item{CNVMouseMoveArcStrokeColor}{Arc stroke color when mouse moving in the element}

\item{CNVMouseMoveArcStrokeWidth}{Arc stroke width when mouse moving in the element}

\item{CNVMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a CNV point anymore.}

\item{CNVMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{CNVMouseOutColor}{Color when mouse moving out the element}

\item{CNVMouseOutArcOpacity}{Arc opacity when mouse moving out the element}

\item{CNVMouseOutArcStrokeColor}{Arc stroke color when mouse moving out the element}

\item{CNVMouseOutArcStrokeWidth}{Arc stroke width when mouse moving out the element}

\item{CNVMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a CNV point.}

\item{CNVMouseUpColor}{Color when mouse moving up the element}

\item{CNVMouseUpArcOpacity}{Arc opacity when mouse clicking the element}

\item{CNVMouseUpArcStrokeColor}{Arc stroke color when mouse clicking the element}

\item{CNVMouseUpArcStrokeWidth}{Arc stroke width when mouse clicking the element}

\item{CNVMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a CNV point.}

\item{CNVMouseOverColor}{Color when mouse moving over the element}

\item{CNVMouseOverArcOpacity}{Arc opacity when mouse moving over the element}

\item{CNVMouseOverArcStrokeColor}{Arc stroke color when mouse moving over the element}

\item{CNVMouseOverArcStrokeWidth}{Arc stroke width when mouse moving over the element}

\item{CNVMouseOverTooltipsSetting}{Default "style1"}

\item{CNVMouseOverTooltipsHtml}{Default " "}

\item{CNVMouseOverTooltipsPosition}{Default "absolute"}

\item{CNVMouseOverTooltipsBackgroundColor}{Default "white"}

\item{CNVMouseOverTooltipsBorderStyle}{Default "solid"}

\item{CNVMouseOverTooltipsBorderWidth}{Default 0}

\item{CNVMouseOverTooltipsPadding}{Default "3px"}

\item{CNVMouseOverTooltipsBorderRadius}{Default "3px"}

\item{CNVMouseOverTooltipsOpacity}{Default 0.8}

\item{HEATMAPMouseEvent}{Default True, open/not open mouse event of HEATMAP module}

\item{HEATMAPMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a HEATMAP point.}

\item{HEATMAPMouseClickColor}{Color when mouse clicking}

\item{HEATMAPMouseClickOpacity}{Opacity when mouse clicking}

\item{HEATMAPMouseClickStrokeColor}{Stroke color when mouse clicking}

\item{HEATMAPMouseClickStrokeWidth}{Stroke width when mouse clicking}

\item{HEATMAPMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a HEATMAP point.}

\item{HEATMAPMouseDownColor}{Color when mouse moving down the element}

\item{HEATMAPMouseDownOpacity}{Opacity when mouse moving down the element}

\item{HEATMAPMouseDownStrokeColor}{Stroke color when mouse moving down the element}

\item{HEATMAPMouseDownStrokeWidth}{Stroke width when mouse moving down the element}

\item{HEATMAPMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a HEATMAP point.}

\item{HEATMAPMouseEnterColor}{Color when mouse entering the element}

\item{HEATMAPMouseEnterOpacity}{Opacity when mouse entering the element}

\item{HEATMAPMouseEnterStrokeColor}{Stroke color when mouse entering the element}

\item{HEATMAPMouseEnterStrokeWidth}{Stroke width when mouse entering the element}

\item{HEATMAPMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a HEATMAP point.}

\item{HEATMAPMouseLeaveColor}{Color when mouse leaving the element}

\item{HEATMAPMouseLeaveOpacity}{Opacity when mouse leaving the element}

\item{HEATMAPMouseLeaveStrokeColor}{Stroke color when mouse leaving the element}

\item{HEATMAPMouseLeaveStrokeWidth}{Stroke width when mouse leaving the element}

\item{HEATMAPMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a HEATMAP point.}

\item{HEATMAPMouseMoveColor}{Color when mouse moving in the element}

\item{HEATMAPMouseMoveOpacity}{Opacity when mouse moving in the element}

\item{HEATMAPMouseMoveStrokeColor}{Stroke color when mouse moving in the element}

\item{HEATMAPMouseMoveStrokeWidth}{Stroke width when mouse moving in the element}

\item{HEATMAPMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a HEATMAP point anymore.}

\item{HEATMAPMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{HEATMAPMouseOutColor}{Color when mouse moving out the element}

\item{HEATMAPMouseOutOpacity}{Opacity when mouse moving out the element}

\item{HEATMAPMouseOutStrokeColor}{Stroke color when mouse moving out the element}

\item{HEATMAPMouseOutStrokeWidth}{Stroke width when mouse moving out the element}

\item{HEATMAPMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a HEATMAP point.}

\item{HEATMAPMouseUpColor}{Color when mouse moving up the element}

\item{HEATMAPMouseUpOpacity}{Opacity when mouse moving up the element}

\item{HEATMAPMouseUpStrokeColor}{Stroke color when mouse moving up the element}

\item{HEATMAPMouseUpStrokeWidth}{Stroke width when mouse moving up the element}

\item{HEATMAPMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a HEATMAP point.}

\item{HEATMAPMouseOverColor}{Color when mouse moving over the element}

\item{HEATMAPMouseOverOpacity}{Opacity when mouse moving over the element}

\item{HEATMAPMouseOverStrokeColor}{Stroke color when mouse moving over the element}

\item{HEATMAPMouseOverStrokeWidth}{Stroke width when mouse moving over the element}

\item{HEATMAPMouseOverTooltipsSetting}{Default "style1"}

\item{HEATMAPMouseOverTooltipsHtml}{Default " "}

\item{HEATMAPMouseOverTooltipsPosition}{Default "absolute"}

\item{HEATMAPMouseOverTooltipsBackgroundColor}{Default "white"}

\item{HEATMAPMouseOverTooltipsBorderStyle}{Default "solid"}

\item{HEATMAPMouseOverTooltipsBorderWidth}{Default 0}

\item{HEATMAPMouseOverTooltipsPadding}{Default "3px"}

\item{HEATMAPMouseOverTooltipsBorderRadius}{Default "3px"}

\item{HEATMAPMouseOverTooltipsOpacity}{Default 0.8}

\item{BUBBLExlink}{Default False, add/not xlink for BUBBLE module}

\item{BUBBLEMouseEvent}{Default True, open/not open mouse event of BUBBLE module}

\item{BUBBLEMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a BUBBLE point.}

\item{BUBBLEMouseClickColor}{Color when mouse clicking}

\item{BUBBLEMouseClickOpacity}{Opacity when mouse clicking}

\item{BUBBLEMouseClickStrokeColor}{Stroke color when mouse clicking}

\item{BUBBLEMouseClickStrokeWidth}{Stroke width when mouse clicking}

\item{BUBBLEMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a BUBBLE point.}

\item{BUBBLEMouseDownColor}{Color when mouse moving down the element}

\item{BUBBLEMouseDownOpacity}{Opacity when mouse moving down the element}

\item{BUBBLEMouseDownStrokeColor}{Stroke color when mouse moving down the element}

\item{BUBBLEMouseDownStrokeWidth}{Stroke width when mouse moving down the element}

\item{BUBBLEMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a BUBBLE point.}

\item{BUBBLEMouseEnterColor}{Color when mouse entering the element}

\item{BUBBLEMouseEnterOpacity}{Opacity when mouse entering the element}

\item{BUBBLEMouseEnterStrokeColor}{Stroke color when mouse entering the element}

\item{BUBBLEMouseEnterStrokeWidth}{Stroke width when mouse entering the element}

\item{BUBBLEMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a BUBBLE point.}

\item{BUBBLEMouseLeaveColor}{Color when mouse leaving the element}

\item{BUBBLEMouseLeaveOpacity}{Opacity when mouse leaving the element}

\item{BUBBLEMouseLeaveStrokeColor}{Stroke color when mouse leaving the element}

\item{BUBBLEMouseLeaveStrokeWidth}{Stroke width when mouse leaving the element}

\item{BUBBLEMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a BUBBLE point.}

\item{BUBBLEMouseMoveColor}{Color when mouse moving in the element}

\item{BUBBLEMouseMoveOpacity}{Opacity when mouse moving in the element}

\item{BUBBLEMouseMoveStrokeColor}{Stroke color when mouse moving in the element}

\item{BUBBLEMouseMoveStrokeWidth}{Stroke width when mouse moving in the element}

\item{BUBBLEMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a BUBBLE point anymore.}

\item{BUBBLEMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{BUBBLEMouseOutColor}{Color when mouse moving out the element}

\item{BUBBLEMouseOutOpacity}{Opacity when mouse moving out the element}

\item{BUBBLEMouseOutStrokeColor}{Stroke color when mouse moving out the element}

\item{BUBBLEMouseOutStrokeWidth}{Stroke width when mouse moving out the element}

\item{BUBBLEMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a BUBBLE point.}

\item{BUBBLEMouseUpColor}{Color when mouse moving up the element}

\item{BUBBLEMouseUpOpacity}{Opacity when mouse moving up the element}

\item{BUBBLEMouseUpStrokeColor}{Stroke color when mouse moving up the element}

\item{BUBBLEMouseUpStrokeWidth}{Stroke width when mouse moving up the element}

\item{BUBBLEMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a BUBBLE point.}

\item{BUBBLEMouseOverColor}{Color when mouse moving over the element}

\item{BUBBLEMouseOverOpacity}{Opacity when mouse moving over the element}

\item{BUBBLEMouseOverStrokeColor}{Stroke color when mouse moving over the element}

\item{BUBBLEMouseOverStrokeWidth}{Stroke width when mouse moving over the element}

\item{BUBBLEMouseOverTooltipsSetting}{Default "style1"}

\item{BUBBLEMouseOverTooltipsHtml}{Default " "}

\item{BUBBLEMouseOverTooltipsPosition}{Default "absolute"}

\item{BUBBLEMouseOverTooltipsBackgroundColor}{Default "white"}

\item{BUBBLEMouseOverTooltipsBorderStyle}{Default "solid"}

\item{BUBBLEMouseOverTooltipsBorderWidth}{Default 0}

\item{BUBBLEMouseOverTooltipsPadding}{Default "3px"}

\item{BUBBLEMouseOverTooltipsBorderRadius}{Default "3px"}

\item{BUBBLEMouseOverTooltipsOpacity}{Default 0.8}

\item{SNPxlink}{Default False, add/not xlink for SNP module}

\item{SNPMouseEvent}{Default True, open/not open mouse event of SNP module}

\item{SNPMouseCombinationEvent}{Default False, open/not COMBINATION module for SNP module}

\item{SNPMouseCombinationImageDisplay}{Defalut False, open/not image display in COMBINATION module for SNP module}

\item{SNPMouseCombinationImageTitle}{Title of the image}

\item{SNPMouseCombinationImageTitleSize, SNPMouseCombinationImageTitleWeight, SNPMouseCombinationImageTitleColor}{Size, weight and color of the title}

\item{SNPMouseCombinationImagePositionX, SNPMouseCombinationImagePositionY}{Coordinates for image}

\item{SNPMouseCombinationImageHeight, SNPMouseCombinationImageWidth}{Height and width of image}

\item{SNPMouseCombinationGraphDisplay}{Defalut False, open/not graph display in COMBINATION module for SNP module}

\item{SNPMouseCombinationGraphTitle}{Title of the graph}

\item{SNPMouseCombinationGraphTitleSize, SNPMouseCombinationGraphTitleWeight, SNPMouseCombinationGraphTitleColor}{Size, weight and color of the title}

\item{SNPMouseCombinationGraphType}{Type of graph}

\item{SNPMouseCombinationGraphPositionX, SNPMouseCombinationGraphPositionY}{Coordinates for graph}

\item{SNPMouseCombinationGraphHeight, SNPMouseCombinationGraphWidth}{Height and width for graph}

\item{SNPMouseCombinationGraphHistogramBarColor}{Bar color of histogram graph}

\item{SNPMouseCombinationGraphHistogramPadding}{Padding between bar of histogram graph}

\item{SNPMouseCombinationGraphHistogramPositionCorrectX}{Correction distance of X axis in histogram}

\item{SNPMouseCombinationGraphPieAutoColor}{Whether use auto color for pie graph or not}

\item{SNPMouseCombinationGraphPieColor}{Color for pie graph if auto color is false}

\item{SNPMouseCombinationGraphPieSize}{Size of pie graph}

\item{SNPMouseCombinationGraphPieStroke}{Whether each pie has a stroke or not}

\item{SNPMouseCombinationGraphPieStrokeColor, SNPMouseCombinationGraphPieStrokeWidth}{The stroke color and width for pie graph}

\item{SNPMouseCombinationGraphPieOpacity}{Opacity for pie graph}

\item{SNPMouseCombinationGraphLineType, SNPMouseCombinationGraphLineColor, SNPMouseCombinationGraphLineWidth}{Line type, color and width for line graph}

\item{SNPMouseCombinationGraphLinePoint}{Whether display the broken point in line graph}

\item{SNPMouseCombinationGraphLinePointSize}{Size of broken point}

\item{SNPMouseCombinationGraphLinePointAutoColor}{Whether display the broken point in auto color}

\item{SNPMouseCombinationGraphLinePointColor}{Color for broken point if auto color is false}

\item{SNPMouseCombinationGraphLinePointStroke}{Whether display the broken point stroke}

\item{SNPMouseCombinationGraphLinePointStrokeColor, SNPMouseCombinationGraphLinePointStrokeWidth}{The stroke color and width for broken point}

\item{SNPMouseCombinationGraphLinePointOpacity}{Opacity for broken line}

\item{SNPMouseCombinationGraphLinePositionCorrectX}{Correction distance of X axis for line}

\item{SNPMouseCombinationTextDisplay}{Defalut False, open/not text display in COMBINATION module for SNP module}

\item{SNPMouseCombinationTextColor, SNPMouseCombinationTextSize, SNPMouseCombinationTextWeight}{The color, size and weight for text}

\item{SNPMouseCombinationTextPositionCorrectX, SNPMouseCombinationTextPositionCorrectY}{The coordinates for text}

\item{SNPMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a SNP point.}

\item{SNPMouseClickColor}{Color after clicking the element}

\item{SNPMouseClickCircleSize}{Circle size after clicking the element}

\item{SNPMouseClickCircleOpacity}{Opacity after clicking the element}

\item{SNPMouseClickCircleStrokeColor}{Stroke color after clicking the element}

\item{SNPMouseClickCircleStrokeWidth}{Stroke width after clicking the element}

\item{SNPMouseClickTextFromData}{First,second,third,fourth column data click to show}

\item{SNPMouseClickTextOpacity}{Text opacity after clicking the element}

\item{SNPMouseClickTextColor}{Text color after clicking the element}

\item{SNPMouseClickTextSize}{Text size after clicking the element}

\item{SNPMouseClickTextPostionX, SNPMouseClickTextPostionY}{Text coordinate after clicking the element}

\item{SNPMouseClickTextDrag}{Whether text is draggable for element}

\item{SNPMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a SNP point.}

\item{SNPMouseDownColor}{Color after mouse moving down the element}

\item{SNPMouseDownCircleSize}{Circle size after mouse moving down the element}

\item{SNPMouseDownCircleOpacity}{Circle opacity after mouse moving down the element}

\item{SNPMouseDownCircleStrokeColor}{Circle stroke color after mouse moving down the element}

\item{SNPMouseDownCircleStrokeWidth}{Circle stroke width after mouse moving down the element}

\item{SNPMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a SNP point.}

\item{SNPMouseEnterColor}{Color after mouse entering enter the element}

\item{SNPMouseEnterCircleSize}{Circle size after mouse entering the element}

\item{SNPMouseEnterCircleOpacity}{Circle opacity after mouse entering the element}

\item{SNPMouseEnterCircleStrokeColor}{Circle stroke color after mouse entering the element}

\item{SNPMouseEnterCircleStrokeWidth}{Circle stroke width after mouse entering the element}

\item{SNPMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a SNP point.}

\item{SNPMouseLeaveColor}{Color after mouse leaving the element}

\item{SNPMouseLeaveCircleSize}{Circle size after mouse leaving the element}

\item{SNPMouseLeaveCircleOpacity}{Circle opacity after mouse leaving the element}

\item{SNPMouseLeaveCircleStrokeColor}{Circle stroke color after mouse leaving the element}

\item{SNPMouseLeaveCircleStrokeWidth}{Circle stroke width after mouse leaving the element}

\item{SNPMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a SNP point.}

\item{SNPMouseMoveColor}{Color after mouse moving in the element}

\item{SNPMouseMoveCircleSize}{Circle size after mouse moving in the element}

\item{SNPMouseMoveCircleOpacity}{Circle opacity after mouse moving in the element}

\item{SNPMouseMoveCircleStrokeColor}{Circle stroke color after mouse moving in the element}

\item{SNPMouseMoveCircleStrokeWidth}{Circle stroke width after mouse moving in the element}

\item{SNPMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a SNP point anymore.}

\item{SNPMouseOutAnimationTime}{Animation time when mouse moving over the element}

\item{SNPMouseOutColor}{Color when mouse moving over the element}

\item{SNPMouseOutCircleSize}{Circle size when mouse moving over the element}

\item{SNPMouseOutCircleOpacity}{Opacity when mouse moving over the element}

\item{SNPMouseOutCircleStrokeColor}{Stroke color when mouse moving over the element}

\item{SNPMouseOutCircleStrokeWidth}{Stroke width when mouse moving over the element}

\item{SNPMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a SNP point.}

\item{SNPMouseUpColor}{Color after mouse moving up the element}

\item{SNPMouseUpCircleSize}{Circle size after mouse moving up the element}

\item{SNPMouseUpCircleOpacity}{Circle opacity after mouse moving up the element}

\item{SNPMouseUpCircleStrokeColor}{Circle stroke color after mouse moving up the element}

\item{SNPMouseUpCircleStrokeWidth}{Circle stroke width after mouse moving up the element}

\item{SNPMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a SNP point.}

\item{SNPMouseOverColor}{Color after mouse moving over the element}

\item{SNPMouseOverCircleSize}{Circle size after mouse moving over the element}

\item{SNPMouseOverCircleOpacity}{Circle opacity after mouse moving over the element}

\item{SNPMouseOverCircleStrokeColor}{Circle stroke color after mouse moving over the element}

\item{SNPMouseOverCircleStrokeWidth}{Circle stroke width after mouse moving over the element}

\item{SNPMouseOverTooltipsSetting}{Default "chr : "}

\item{SNPMouseOverTooltipsHtml}{Default " "}

\item{SNPMouseOverTooltipsPosition}{Position for tooltips when mouse moving over}

\item{SNPMouseOverTooltipsBackgroundColor}{Background color for tooltips when mouse moving over}

\item{SNPMouseOverTooltipsBorderStyle}{Border style for tooltips when mouse moving over}

\item{SNPMouseOverTooltipsBorderWidth}{Border width for tooltips when mouse moving over}

\item{SNPMouseOverTooltipsPadding}{Padding for tooltips when mouse moving over}

\item{SNPMouseOverTooltipsBorderRadius}{Border radius for tooltips when mouse moving over}

\item{SNPMouseOverTooltipsOpacity}{Opacity for tooltips when mouse moving over}

\item{LINKxlink}{Default False, add/not xlink for LINK module}

\item{LINKMouseEvent}{Default True, open/not open mouse event of LINK module}

\item{LINKMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a LINK point.}

\item{LINKMouseClickOpacity}{Opacity when mouse clicking}

\item{LINKMouseClickStrokeColor}{Stroke color when mouse clicking}

\item{LINKMouseClickStrokeWidth}{Stroke width when mouse clicking}

\item{LINKMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a LINK point.}

\item{LINKMouseDownOpacity}{Opacity when mouse moving down the element}

\item{LINKMouseDownStrokeColor}{Stroke color when mouse moving down the element}

\item{LINKMouseDownStrokeWidth}{Stroke width when mouse moving down the element}

\item{LINKMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a LINK point.}

\item{LINKMouseEnterOpacity}{Opacity when mouse entering the element}

\item{LINKMouseEnterStrokeColor}{Stroke color when mouse entering the element}

\item{LINKMouseEnterStrokeWidth}{Stroke width when mouse entering the element}

\item{LINKMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a LINK point.}

\item{LINKMouseLeaveOpacity}{Opacity when mouse leaving the element}

\item{LINKMouseLeaveStrokeColor}{Stroke color when mouse leaving the element}

\item{LINKMouseLeaveStrokeWidth}{Stroke width when mouse leaving the element}

\item{LINKMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a LINK point.}

\item{LINKMouseMoveOpacity}{Opacity when mouse moving in the element}

\item{LINKMouseMoveStrokeColor}{Stroke color when mouse moving in the element}

\item{LINKMouseMoveStrokeWidth}{Stroke width when mouse moving in the element}

\item{LINKMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a LINK point anymore.}

\item{LINKMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{LINKMouseOutOpacity}{Opacity when mouse moving out the element}

\item{LINKMouseOutStrokeColor}{Stroke color when mouse moving out the element}

\item{LINKMouseOutStrokeWidth}{Stroke width when mouse moving out the element}

\item{LINKMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a LINK point.}

\item{LINKMouseUpOpacity}{Opacity when mouse moving up the element}

\item{LINKMouseUpStrokeColor}{Stroke color when mouse moving up the element}

\item{LINKMouseUpStrokeWidth}{Stroke width when mouse moving up the element}

\item{LINKMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a LINK point.}

\item{LINKMouseOverOpacity}{Opacity when mouse moving over the element}

\item{LINKMouseOverStrokeColor}{Stroke color when mouse moving over the element}

\item{LINKMouseOverStrokeWidth}{Stroke width when mouse moving over the element}

\item{LINKMouseOverTooltipsSetting}{Default "style1"}

\item{LINKMouseOverTooltipsHtml}{Default " "}

\item{LINKMouseOverTooltipsPosition}{Default "absolute"}

\item{LINKMouseOverTooltipsBackgroundColor}{Default "white"}

\item{LINKMouseOverTooltipsBorderStyle}{Default "solid"}

\item{LINKMouseOverTooltipsBorderWidth}{Default 0}

\item{LINKMouseOverTooltipsPadding}{Default "3px"}

\item{LINKMouseOverTooltipsBorderRadius}{Default "3px"}

\item{LINKMouseOverTooltipsOpacity}{Default 0.8}

\item{LINKLabelDragEvent}{Defalut False, draggable for the label of LINK module}

\item{CHORDMouseEvent}{Default True, open/not open mouse event of CHORD module}

\item{CHORDMouseFillColorExcluded}{A type of color in character, chord in this color will be hided}

\item{CHORDMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a CHORD point.}

\item{CHORDMouseClickOpacity}{Opacity when mouse clicking}

\item{CHORDMouseClickStrokeColor}{Stroke color when mouse clicking}

\item{CHORDMouseClickStrokeWidth}{Stroke width when mouse clicking}

\item{CHORDMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a CHORD point.}

\item{CHORDMouseDownOpacity}{Opacity when mouse moving down the element}

\item{CHORDMouseDownStrokeColor}{Stroke color when mouse moving down the element}

\item{CHORDMouseDownStrokeWidth}{Stroke width when mouse moving down the element}

\item{CHORDMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a CHORD point.}

\item{CHORDMouseEnterOpacity}{Opacity when mouse entering the element}

\item{CHORDMouseEnterStrokeColor}{Stroke color when mouse entering the element}

\item{CHORDMouseEnterStrokeWidth}{Stroke width when mouse entering the element}

\item{CHORDMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a CHORD point.}

\item{CHORDMouseLeaveOpacity}{Opacity when mouse leaving the element}

\item{CHORDMouseLeaveStrokeColor}{Stroke color when mouse leaving the element}

\item{CHORDMouseLeaveStrokeWidth}{Stroke width when mouse leaving the element}

\item{CHORDMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a CHORD point.}

\item{CHORDMouseMoveOpacity}{Opacity when mouse moving in the element}

\item{CHORDMouseMoveStrokeColor}{Stroke color when mouse moving in the element}

\item{CHORDMouseMoveStrokeWidth}{Stroke width when mouse moving in the element}

\item{CHORDMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a CHORD point anymore.}

\item{CHORDMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{CHORDMouseOutOpacity}{Opacity when mouse moving out the element}

\item{CHORDMouseOutStrokeColor}{Stroke color when mouse moving out the element}

\item{CHORDMouseOutStrokeWidth}{Stroke width when mouse moving out the element}

\item{CHORDMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a CHORD point.}

\item{CHORDMouseUpOpacity}{Opacity when mouse moving up the element}

\item{CHORDMouseUpStrokeColor}{Stroke color when mouse moving up the element}

\item{CHORDMouseUpStrokeWidth}{Stroke width when mouse moving up the element}

\item{CHORDMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a CHORD point.}

\item{CHORDMouseOverOpacity}{Opacity when mouse moving over the element}

\item{CHORDMouseOverStrokeColor}{Stroke color when mouse moving over the element}

\item{CHORDMouseOverStrokeWidth}{Stroke width when mouse moving over the element}

\item{HISTOGRAMxlink}{Default False, add/not xlink for HISTOGRAM module}

\item{HISTOGRAMMouseEvent}{Default True, open/not open mouse event of HISTOGRAM module}

\item{HISTOGRAMMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a HISTOGRAM point.}

\item{HISTOGRAMMouseClickColor}{Color when mouse clicking}

\item{HISTOGRAMMouseClickOpacity}{Opacity when mouse clicking}

\item{HISTOGRAMMouseClickStrokeColor}{Stroke color when mouse clicking}

\item{HISTOGRAMMouseClickStrokeWidth}{Stroke width when mouse clicking}

\item{HISTOGRAMMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a HISTOGRAM point.}

\item{HISTOGRAMMouseDownColor}{Color when mouse moving down the element}

\item{HISTOGRAMMouseDownOpacity}{Opacity when mouse moving up the element}

\item{HISTOGRAMMouseDownStrokeColor}{Stroke color when mouse moving up the element}

\item{HISTOGRAMMouseDownStrokeWidth}{Stroke width when mouse moving up the element}

\item{HISTOGRAMMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a HISTOGRAM point.}

\item{HISTOGRAMMouseEnterColor}{Color when mouse entering the element}

\item{HISTOGRAMMouseEnterOpacity}{Opacity when mouse entering the element}

\item{HISTOGRAMMouseEnterStrokeColor}{Stroke color when mouse entering the element}

\item{HISTOGRAMMouseEnterStrokeWidth}{Stroke width when mouse entering the element}

\item{HISTOGRAMMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a HISTOGRAM point.}

\item{HISTOGRAMMouseLeaveColor}{Color when mouse leaving the element}

\item{HISTOGRAMMouseLeaveOpacity}{Opacity when mouse leaving the element}

\item{HISTOGRAMMouseLeaveStrokeColor}{Stroke color when mouse leaving the element}

\item{HISTOGRAMMouseLeaveStrokeWidth}{Stroke width when mouse leaving the element}

\item{HISTOGRAMMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a HISTOGRAM point.}

\item{HISTOGRAMMouseMoveColor}{Color when mouse moving in the element}

\item{HISTOGRAMMouseMoveOpacity}{Opacity when mouse moving in the element}

\item{HISTOGRAMMouseMoveStrokeColor}{Stroke color when mouse moving in the element}

\item{HISTOGRAMMouseMoveStrokeWidth}{Stroke width when mouse moving in the element}

\item{HISTOGRAMMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a HISTOGRAM point anymore.}

\item{HISTOGRAMMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{HISTOGRAMMouseOutColor}{Color when mouse moving out the element}

\item{HISTOGRAMMouseOutOpacity}{Opacity when mouse moving out the element}

\item{HISTOGRAMMouseOutStrokeColor}{Stroke color when mouse moving out the element}

\item{HISTOGRAMMouseOutStrokeWidth}{Stroke width when mouse moving out the element}

\item{HISTOGRAMMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a HISTOGRAM point.}

\item{HISTOGRAMMouseUpColor}{Color when mouse moving up the element}

\item{HISTOGRAMMouseUpOpacity}{Opacity when mouse moving up the element}

\item{HISTOGRAMMouseUpStrokeColor}{Stroke color when mouse moving up the element}

\item{HISTOGRAMMouseUpStrokeWidth}{Stroke width when mouse moving up the element}

\item{HISTOGRAMMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a HISTOGRAM point.}

\item{HISTOGRAMMouseOverColor}{Color when mouse moving over the element}

\item{HISTOGRAMMouseOverOpacity}{Opacity when mouse moving over the element}

\item{HISTOGRAMMouseOverStrokeColor}{Stroke color when mouse moving over the element}

\item{HISTOGRAMMouseOverStrokeWidth}{Stroke width when mouse moving over the element}

\item{HISTOGRAMMouseOverTooltipsSetting}{Default "style1"}

\item{HISTOGRAMMouseOverTooltipsHtml}{Default " "}

\item{HISTOGRAMMouseOverTooltipsPosition}{Default "absolute"}

\item{HISTOGRAMMouseOverTooltipsBackgroundColor}{Default "white"}

\item{HISTOGRAMMouseOverTooltipsBorderStyle}{Default "solid"}

\item{HISTOGRAMMouseOverTooltipsBorderWidth}{Default 0}

\item{HISTOGRAMMouseOverTooltipsPadding}{Default "3px"}

\item{HISTOGRAMMouseOverTooltipsBorderRadius}{Default "3px"}

\item{HISTOGRAMMouseOverTooltipsOpacity}{Default 0.8}

\item{LINEMouseEvent}{Default True, open/not open mouse event of LINE module}

\item{LINEMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a LINE point.}

\item{LINEMouseClickLineOpacity}{Line opacity when mouse clicking the element}

\item{LINEMouseClickLineStrokeColor}{Stroke color when mouse clicking the element}

\item{LINEMouseClickLineStrokeWidth}{Stroke width when mouse clicking the element}

\item{LINEMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a LINE point.}

\item{LINEMouseDownLineOpacity}{Line opacity when mouse moving down the element}

\item{LINEMouseDownLineStrokeColor}{Stroke color when mouse moving down the element}

\item{LINEMouseDownLineStrokeWidth}{Stroke width when mouse moving down the element}

\item{LINEMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a LINE point.}

\item{LINEMouseEnterLineOpacity}{Line opacity when mouse entering the element}

\item{LINEMouseEnterLineStrokeColor}{Stroke color when mouse entering the element}

\item{LINEMouseEnterLineStrokeWidth}{Stroke width when mouse entering the element}

\item{LINEMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a LINE point.}

\item{LINEMouseLeaveLineOpacity}{Line opacity when mouse leaving the element}

\item{LINEMouseLeaveLineStrokeColor}{Stroke color when mouse leaving the element}

\item{LINEMouseLeaveLineStrokeWidth}{Stroke width when mouse leaving the element}

\item{LINEMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a LINE point.}

\item{LINEMouseMoveLineOpacity}{Line opacity when mouse moving in the element}

\item{LINEMouseMoveLineStrokeColor}{Stroke color when mouse moving in the element}

\item{LINEMouseMoveLineStrokeWidth}{Stroke width when mouse moving in the element}

\item{LINEMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a LINE point anymore.}

\item{LINEMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{LINEMouseOutLineOpacity}{Line opacity when mouse moving out the element}

\item{LINEMouseOutLineStrokeColor}{Stroke color when mouse moving out the element}

\item{LINEMouseOutLineStrokeWidth}{Stroke width when mouse moving out the element}

\item{LINEMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a LINE point.}

\item{LINEMouseUpLineOpacity}{Line opacity when mouse moving up the element}

\item{LINEMouseUpLineStrokeColor}{Stroke color when mouse moving up the element}

\item{LINEMouseUpLineStrokeWidth}{Stroke width when mouse moving up the element}

\item{LINEMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a LINE point.}

\item{LINEMouseOverLineOpacity}{Line opacity when mouse moving over the element}

\item{LINEMouseOverLineStrokeColor}{Stroke color when mouse moving over the element}

\item{LINEMouseOverLineStrokeWidth}{Stroke width when mouse moving over the element}

\item{LINEMouseOverTooltipsSetting}{Default "style1"}

\item{LINEMouseOverTooltipsHtml}{Default " "}

\item{LINEMouseOverTooltipsPosition}{Default "absolute"}

\item{LINEMouseOverTooltipsBackgroundColor}{Default "white"}

\item{LINEMouseOverTooltipsBorderStyle}{Default "solid"}

\item{LINEMouseOverTooltipsBorderWidth}{Default 0}

\item{LINEMouseOverTooltipsPadding}{Default "3px"}

\item{LINEMouseOverTooltipsBorderRadius}{Default "3px"}

\item{LINEMouseOverTooltipsOpacity}{Default 0.8}

\item{WIGMouseEvent}{Default True, open/not open mouse event of WIG module}

\item{WIGMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a WIG point.}

\item{WIGMouseClickLineOpacity}{Line opacity when mouse clicking the element}

\item{WIGMouseClickLineStrokeColor}{Stroke color when mouse clicking the element}

\item{WIGMouseClickLineStrokeWidth}{Stroke width when mouse clicking the element}

\item{WIGMouseClickFillColor}{Filling color when mouse clicking the element}

\item{WIGMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a WIG point.}

\item{WIGMouseDownLineOpacity}{Line opacity when mouse moving down the element}

\item{WIGMouseDownLineStrokeColor}{Stroke color when mouse moving down the element}

\item{WIGMouseDownLineStrokeWidth}{Stroke width when mouse moving down the element}

\item{WIGMouseDownFillColor}{Filling color when mouse moving down the element}

\item{WIGMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a WIG point.}

\item{WIGMouseEnterLineOpacity}{Line opacity when mouse entering the element}

\item{WIGMouseEnterLineStrokeColor}{Stroke color when mouse entering the element}

\item{WIGMouseEnterLineStrokeWidth}{Stroke width when mouse entering the element}

\item{WIGMouseEnterFillColor}{Filling color when mouse entering the element}

\item{WIGMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a WIG point.}

\item{WIGMouseLeaveLineOpacity}{Line opacity when mouse leaving the element}

\item{WIGMouseLeaveLineStrokeColor}{Stroke color when mouse leaving the element}

\item{WIGMouseLeaveLineStrokeWidth}{Stroke width when mouse leaving the element}

\item{WIGMouseLeaveFillColor}{Filling color when mouse leaving the element}

\item{WIGMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a WIG point.}

\item{WIGMouseMoveLineOpacity}{Line opacity when mouse moving in the element}

\item{WIGMouseMoveLineStrokeColor}{Stroke color when mouse moving in the element}

\item{WIGMouseMoveLineStrokeWidth}{Stroke width when mouse moving in the element}

\item{WIGMouseMoveFillColor}{Filling color when mouse leaving the element}

\item{WIGMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a WIG point anymore.}

\item{WIGMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{WIGMouseOutLineOpacity}{Line opacity when mouse moving out the element}

\item{WIGMouseOutLineStrokeColor}{Stroke color when mouse moving out the element}

\item{WIGMouseOutLineStrokeWidth}{Stroke width when mouse moving out the element}

\item{WIGMouseOutFillColor}{Filling color when mouse moving out the element}

\item{WIGMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a WIG point.}

\item{WIGMouseUpLineOpacity}{Line opacity when mouse moving up the element}

\item{WIGMouseUpLineStrokeColor}{Stroke color when mouse moving up the element}

\item{WIGMouseUpLineStrokeWidth}{Stroke width when mouse moving up the element}

\item{WIGMouseUpFillColor}{Filling color when mouse moving up the element}

\item{WIGMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a WIG point.}

\item{WIGMouseOverLineOpacity}{Line opacity when mouse moving over the element}

\item{WIGMouseOverLineStrokeColor}{Stroke color when mouse moving over the element}

\item{WIGMouseOverLineStrokeWidth}{Stroke width when mouse moving over the element}

\item{WIGMouseOverFillColor}{Filling color when mouse moving over the element}

\item{WIGMouseOverTooltipsSetting}{Default "style1"}

\item{WIGMouseOverTooltipsHtml}{Default " "}

\item{WIGMouseOverTooltipsPosition}{Default "absolute"}

\item{WIGMouseOverTooltipsBackgroundColor}{Default "white"}

\item{WIGMouseOverTooltipsBorderStyle}{Default "solid"}

\item{WIGMouseOverTooltipsBorderWidth}{Default 0}

\item{WIGMouseOverTooltipsPadding}{Default "3px"}

\item{WIGMouseOverTooltipsBorderRadius}{Default "3px"}

\item{WIGMouseOverTooltipsOpacity}{Default 0.8}

\item{SCATTERxlink}{Default False, add/not xlink for SCATTER module}

\item{SCATTERMouseEvent}{Default True, open/not open mouse event of SCATTER module}

\item{SCATTERMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a SCATTER point.}

\item{SCATTERMouseClickColor}{Color when mouse clicking  the element}

\item{SCATTERMouseClickCircleSize}{Circle size when mouse clicking the element}

\item{SCATTERMouseClickCircleOpacity}{Circle opacity when mouse clicking the element}

\item{SCATTERMouseClickCircleStrokeColor}{Circle stroke color when mouse clicking the element}

\item{SCATTERMouseClickCircleStrokeWidth}{Circle stroke width when mouse clicking the element}

\item{SCATTERMouseClickTextFromData}{Text column when mouse clicking the element}

\item{SCATTERMouseClickTextOpacity}{Text opacity when mouse clicking the element}

\item{SCATTERMouseClickTextColor}{Text color when mouse clicking the element}

\item{SCATTERMouseClickTextSize}{Text size when mouse clicking the element}

\item{SCATTERMouseClickTextPostionX, SCATTERMouseClickTextPostionY}{Text coordinates when mouse clicking the element}

\item{SCATTERMouseClickTextDrag}{Whether text is draggable when clicing element}

\item{SCATTERMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a SCATTER point.}

\item{SCATTERMouseDownColor}{Color when mouse moving down the element}

\item{SCATTERMouseDownCircleSize}{Circle size when mouse moving down the element}

\item{SCATTERMouseDownCircleOpacity}{Circle opacity when mouse moving down the element}

\item{SCATTERMouseDownCircleStrokeColor}{Circle stroke color when mouse moving down the element}

\item{SCATTERMouseDownCircleStrokeWidth}{Circle stroke width when mouse moving down the element}

\item{SCATTERMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a SCATTER point.}

\item{SCATTERMouseEnterColor}{Color when mouse entering the element}

\item{SCATTERMouseEnterCircleSize}{Circle size when mouse entering the element}

\item{SCATTERMouseEnterCircleOpacity}{Circle opacity when mouse entering the element}

\item{SCATTERMouseEnterCircleStrokeColor}{Circle stroke color when mouse entering the element}

\item{SCATTERMouseEnterCircleStrokeWidth}{Circle stroke width when mouse entering the element}

\item{SCATTERMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a SCATTER point.}

\item{SCATTERMouseLeaveColor}{Color when mouse leaving the element}

\item{SCATTERMouseLeaveCircleSize}{Circle size when mouse leaving the element}

\item{SCATTERMouseLeaveCircleOpacity}{Circle opacity when mouse leaving the element}

\item{SCATTERMouseLeaveCircleStrokeColor}{Circle stroke color when mouse leaving the element}

\item{SCATTERMouseLeaveCircleStrokeWidth}{Circle stroke width when mouse leaving the element}

\item{SCATTERMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a SCATTER point.}

\item{SCATTERMouseMoveColor}{Color when mouse moving in the element}

\item{SCATTERMouseMoveCircleSize}{Circle size when mouse moving in the element}

\item{SCATTERMouseMoveCircleOpacity}{Circle opacity when mouse moving in the element}

\item{SCATTERMouseMoveCircleStrokeColor}{Circle stroke color when mouse moving in the element}

\item{SCATTERMouseMoveCircleStrokeWidth}{Circle stroke width when mouse moving in the element}

\item{SCATTERMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a SCATTER point anymore.}

\item{SCATTERMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{SCATTERMouseOutColor}{Color when mouse moving out the element}

\item{SCATTERMouseOutCircleSize}{Circle size when mouse moving out the element}

\item{SCATTERMouseOutCircleOpacity}{Circle opacity when mouse moving out the element}

\item{SCATTERMouseOutCircleStrokeColor}{Circle stroke color when mouse moving out the element}

\item{SCATTERMouseOutCircleStrokeWidth}{Circle stroke width when mouse moving out the element}

\item{SCATTERMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a SCATTER point.}

\item{SCATTERMouseUpColor}{Color when mouse moving up the element}

\item{SCATTERMouseUpCircleSize}{Circle size when mouse moving up the element}

\item{SCATTERMouseUpCircleOpacity}{Circle opacity when mouse moving up the element}

\item{SCATTERMouseUpCircleStrokeColor}{Circle stroke color when mouse moving up the element}

\item{SCATTERMouseUpCircleStrokeWidth}{Circle stroke width when mouse moving up the element}

\item{SCATTERMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a SCATTER point.}

\item{SCATTERMouseOverColor}{Color when mouse moving over the element}

\item{SCATTERMouseOverCircleSize}{Circle size when mouse moving over the element}

\item{SCATTERMouseOverCircleOpacity}{Circle opacity when mouse moving over the element}

\item{SCATTERMouseOverCircleStrokeColor}{Circle stroke color when mouse moving over the element}

\item{SCATTERMouseOverCircleStrokeWidth}{Circle stroke width when mouse moving over the element}

\item{SCATTERMouseOverTooltipsSetting}{Default "style1"}

\item{SCATTERMouseOverTooltipsHtml}{Default " "}

\item{SCATTERMouseOverTooltipsPosition}{Default "absolute"}

\item{SCATTERMouseOverTooltipsBackgroundColor}{Default "white"}

\item{SCATTERMouseOverTooltipsBorderStyle}{Default "solid"}

\item{SCATTERMouseOverTooltipsBorderWidth}{Default 0}

\item{SCATTERMouseOverTooltipsPadding}{Default "3px"}

\item{SCATTERMouseOverTooltipsBorderRadius}{Default "3px"}

\item{SCATTERMouseOverTooltipsOpacity}{Default 0.8}

\item{ARCxlink}{Default False, add/not xlink for ARC module}

\item{ARCMouseEvent}{Default True, open/not open mouse event of ARC module}

\item{ARCMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a ARC point.}

\item{ARCMouseClickColor}{Color when mouse clicking the element}

\item{ARCMouseClickArcOpacity}{Arc opacity when mouse clicking the element}

\item{ARCMouseClickArcStrokeColor}{Arc stroke color when mouse clicking the element}

\item{ARCMouseClickArcStrokeWidth}{Arc stroke width when mouse clicking the element}

\item{ARCMouseClickTextFromData}{Text column when mouse clicking the element}

\item{ARCMouseClickTextOpacity}{Text opacity when mouse clicking the element}

\item{ARCMouseClickTextColor}{Text color when mouse clicking the element}

\item{ARCMouseClickTextSize}{Text size when mouse clicking the element}

\item{ARCMouseClickTextPostionX, ARCMouseClickTextPostionY}{Text coordinates when mouse clicking the element}

\item{ARCMouseClickTextDrag}{Whether text is draggable when mouse clicking the element}

\item{ARCMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a ARC point.}

\item{ARCMouseDownColor}{Color when mouse moving down the element}

\item{ARCMouseDownArcOpacity}{Arc opacity when mouse moving down the element}

\item{ARCMouseDownArcStrokeColor}{Arc stroke color when mouse moving down the element}

\item{ARCMouseDownArcStrokeWidth}{Arc stroke width when mouse moving down the element}

\item{ARCMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a ARC point.}

\item{ARCMouseEnterColor}{Color when mouse entering the element}

\item{ARCMouseEnterArcOpacity}{Arc opacity when mouse entering the element}

\item{ARCMouseEnterArcStrokeColor}{Arc stroke color when mouse entering the element}

\item{ARCMouseEnterArcStrokeWidth}{Arc stroke width when mouse entering the element}

\item{ARCMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a ARC point.}

\item{ARCMouseLeaveColor}{Color when mouse leaving the element}

\item{ARCMouseLeaveArcOpacity}{Arc opacity when mouse leaving the element}

\item{ARCMouseLeaveArcStrokeColor}{Arc stroke color when mouse leaving the element}

\item{ARCMouseLeaveArcStrokeWidth}{Arc stroke width when mouse leaving the element}

\item{ARCMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a ARC point.}

\item{ARCMouseMoveColor}{Color when mouse moving in the element}

\item{ARCMouseMoveArcOpacity}{Arc opacity when mouse moving in the element}

\item{ARCMouseMoveArcStrokeColor}{Arc stroke color when mouse moving in the element}

\item{ARCMouseMoveArcStrokeWidth}{Arc stroke width when mouse moving in the element}

\item{ARCMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a ARC point anymore.}

\item{ARCMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{ARCMouseOutColor}{Color when mouse moving out the element}

\item{ARCMouseOutArcOpacity}{Arc opacity when mouse moving out the element}

\item{ARCMouseOutArcStrokeColor}{Arc stroke color when mouse moving out the element}

\item{ARCMouseOutArcStrokeWidth}{Arc stroke width when mouse moving out the element}

\item{ARCMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a ARC point.}

\item{ARCMouseUpColor}{Color when mouse moving up the element}

\item{ARCMouseUpArcOpacity}{Arc opacity when mouse moving up the element}

\item{ARCMouseUpArcStrokeColor}{Arc stroke color when mouse moving up the element}

\item{ARCMouseUpArcStrokeWidth}{Arc stroke width when mouse moving up the element}

\item{ARCMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a ARC point.}

\item{ARCMouseOverColor}{Color when mouse moving over the element}

\item{ARCMouseOverArcOpacity}{Arc opacity when mouse moving over the element}

\item{ARCMouseOverArcStrokeColor}{Arc stroke color when mouse moving over the element}

\item{ARCMouseOverArcStrokeWidth}{Arc stroke width when mouse moving over the element}

\item{ARCMouseOverTooltipsSetting}{Default "style1"}

\item{ARCMouseOverTooltipsHtml}{Default " "}

\item{ARCMouseOverTooltipsPosition}{Default "absolute"}

\item{ARCMouseOverTooltipsBackgroundColor}{Default "white"}

\item{ARCMouseOverTooltipsBorderStyle}{Default "solid"}

\item{ARCMouseOverTooltipsBorderWidth}{Default 0}

\item{ARCMouseOverTooltipsPadding}{Default "3px"}

\item{ARCMouseOverTooltipsBorderRadius}{Default "3px"}

\item{ARCMouseOverTooltipsOpacity}{Default 0.8}

\item{GENExlink}{Default False, add/not xlink for GENE module}

\item{GENEMouseEvent}{Default True, open/not open mouse event of GENE module}

\item{GENEMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a GENE point.}

\item{GENEMouseClickColor}{Color when mouse clicking the element}

\item{GENEMouseClickArcOpacity}{Arc opacity when mouse clicking the element}

\item{GENEMouseClickArcStrokeColor}{Arc stroke color when mouse clicking the element}

\item{GENEMouseClickArcStrokeWidth}{Arc stroke width when mouse clicking the element}

\item{GENEMouseClickTextFromData}{Text column when mouse clicking the element}

\item{GENEMouseClickTextOpacity}{Text opacity when mouse clicking the element}

\item{GENEMouseClickTextColor}{Text color when mouse clicking the element}

\item{GENEMouseClickTextSize}{Text size when mouse clicking the element}

\item{GENEMouseClickTextPostionX, GENEMouseClickTextPostionY}{Text coordinates when mouse clicking the element}

\item{GENEMouseClickTextDrag}{Whether text is draggable when mouse clicking the element}

\item{GENEMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a GENE point.}

\item{GENEMouseDownColor}{Color when mouse moving down the element}

\item{GENEMouseDownArcOpacity}{Arc opacity when mouse moving down the element}

\item{GENEMouseDownArcStrokeColor}{Arc stroke color when mouse moving down the element}

\item{GENEMouseDownArcStrokeWidth}{Arc stroke width when mouse moving down the element}

\item{GENEMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a GENE point.}

\item{GENEMouseEnterColor}{Color when mouse entering the element}

\item{GENEMouseEnterArcOpacity}{Arc opacity when mouse entering the element}

\item{GENEMouseEnterArcStrokeColor}{Arc stroke color when mouse entering the element}

\item{GENEMouseEnterArcStrokeWidth}{Arc stroke width when mouse entering the element}

\item{GENEMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a GENE point.}

\item{GENEMouseLeaveColor}{Color when mouse leaving the element}

\item{GENEMouseLeaveArcOpacity}{Arc opacity when mouse leaving the element}

\item{GENEMouseLeaveArcStrokeColor}{Arc stroke color when mouse leaving the element}

\item{GENEMouseLeaveArcStrokeWidth}{Arc stroke width when mouse leaving the element}

\item{GENEMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a GENE point.}

\item{GENEMouseMoveColor}{Color when mouse moving in the element}

\item{GENEMouseMoveArcOpacity}{Arc opacity when mouse moving in the element}

\item{GENEMouseMoveArcStrokeColor}{Arc stroke color when mouse moving in the element}

\item{GENEMouseMoveArcStrokeWidth}{Arc stroke width when mouse moving in the element}

\item{GENEMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a GENE point anymore.}

\item{GENEMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{GENEMouseOutColor}{Color when mouse moving out the element}

\item{GENEMouseOutArcOpacity}{Arc opacity when mouse moving out the element}

\item{GENEMouseOutArcStrokeColor}{Arc stroke color when mouse moving out the element}

\item{GENEMouseOutArcStrokeWidth}{Arc stroke width when mouse moving out the element}

\item{GENEMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a GENE point.}

\item{GENEMouseUpColor}{Color when mouse moving up the element}

\item{GENEMouseUpArcOpacity}{Arc opacity when mouse moving up the element}

\item{GENEMouseUpArcStrokeColor}{Arc stroke color when mouse moving up the element}

\item{GENEMouseUpArcStrokeWidth}{Arc stroke width when mouse moving up the element}

\item{GENEMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a GENE point.}

\item{GENEMouseOverColor}{Color when mouse moving over the element}

\item{GENEMouseOverArcOpacity}{Arc opacity when mouse moving over the element}

\item{GENEMouseOverArcStrokeColor}{Arc stroke color when mouse moving over the element}

\item{GENEMouseOverArcStrokeWidth}{Arc stroke width when mouse moving over the element}

\item{GENEMouseOverTooltipsSetting}{Default "style1"}

\item{GENEMouseOverTooltipsHtml}{Default " "}

\item{GENEMouseOverTooltipsPosition}{Default "absolute"}

\item{GENEMouseOverTooltipsBackgroundColor}{Default "white"}

\item{GENEMouseOverTooltipsBorderStyle}{Default "solid"}

\item{GENEMouseOverTooltipsBorderWidth}{Default 0}

\item{GENEMouseOverTooltipsPadding}{Default "3px"}

\item{GENEMouseOverTooltipsBorderRadius}{Default "3px"}

\item{GENEMouseOverTooltipsOpacity}{Default 0.8}

\item{LOLLIPOPxlink}{Default False, add/not xlink for LOLLIPOP module}

\item{LOLLIPOPMouseEvent}{Default True, open/not open mouse event of LOLLIPOP module}

\item{LOLLIPOPMouseClickDisplay}{Default False, show/not the tooltip when mouse click on a LOLLIPOP point.}

\item{LOLLIPOPMouseClickColor}{Color when mouse clicking}

\item{LOLLIPOPMouseClickCircleSize}{Circle size when mouse clicking the element}

\item{LOLLIPOPMouseClickCircleOpacity}{Circle opacity when mouse clicking the element}

\item{LOLLIPOPMouseClickCircleStrokeColor}{Circle stroke color when mouse clicking the element}

\item{LOLLIPOPMouseClickCircleStrokeWidth}{Circle stroke width when mouse clicking the element}

\item{LOLLIPOPMouseClickTextFromData}{Text column when mouse clicking the element}

\item{LOLLIPOPMouseClickTextOpacity}{Text opacity when mouse clicking the element}

\item{LOLLIPOPMouseClickTextColor}{Text color when mouse clicking the element}

\item{LOLLIPOPMouseClickTextSize}{Text size when mouse clicking the element}

\item{LOLLIPOPMouseClickTextPostionX, LOLLIPOPMouseClickTextPostionY}{Text coordinates when mouse clicking the element}

\item{LOLLIPOPMouseClickTextDrag}{Whether text is draggable when mouse clicking the element}

\item{LOLLIPOPMouseDownDisplay}{Default False, show/not the tooltip when mouse click down a LOLLIPOP point.}

\item{LOLLIPOPMouseDownColor}{Color when mouse moving down the element}

\item{LOLLIPOPMouseDownCircleSize}{Circle size when mouse moving down the element}

\item{LOLLIPOPMouseDownCircleOpacity}{Circle opacity when mouse moving down the element}

\item{LOLLIPOPMouseDownCircleStrokeColor}{Circle stroke color when mouse moving down the element}

\item{LOLLIPOPMouseDownCircleStrokeWidth}{Circle stroke width when mouse moving down the element}

\item{LOLLIPOPMouseEnterDisplay}{Default False, show/not the tooltip when mouse mover over a LOLLIPOP point.}

\item{LOLLIPOPMouseEnterColor}{Color when mouse entering the element}

\item{LOLLIPOPMouseEnterCircleSize}{Circle size when mouse entering the element}

\item{LOLLIPOPMouseEnterCircleOpacity}{Circle opacity when mouse entering the element}

\item{LOLLIPOPMouseEnterCircleStrokeColor}{Circle stroke color when mouse entering the element}

\item{LOLLIPOPMouseEnterCircleStrokeWidth}{Circle stroke width when mouse entering the element}

\item{LOLLIPOPMouseLeaveDisplay}{Default False, show/not the tooltip when mouse mover leave a LOLLIPOP point.}

\item{LOLLIPOPMouseLeaveColor}{Color when mouse leaving the element}

\item{LOLLIPOPMouseLeaveCircleSize}{Circle size when mouse leaving the element}

\item{LOLLIPOPMouseLeaveCircleOpacity}{Circle opacity when mouse leaving the element}

\item{LOLLIPOPMouseLeaveCircleStrokeColor}{Circle stroke color when mouse leaving the element}

\item{LOLLIPOPMouseLeaveCircleStrokeWidth}{Circle stroke width when mouse leaving the element}

\item{LOLLIPOPMouseMoveDisplay}{Default False, show/not the tooltip when mouse move into a LOLLIPOP point.}

\item{LOLLIPOPMouseMoveColor}{Color when mouse moving in the element}

\item{LOLLIPOPMouseMoveCircleSize}{Circle size when mouse moving in the element}

\item{LOLLIPOPMouseMoveCircleOpacity}{Circle opacity when mouse moving in the element}

\item{LOLLIPOPMouseMoveCircleStrokeColor}{Circle stroke color when mouse moving in the element}

\item{LOLLIPOPMouseMoveCircleStrokeWidth}{Circle stroke width when mouse moving in the element}

\item{LOLLIPOPMouseOutDisplay}{Defalut False, hide/not tooltip when mouse is not hovering a LOLLIPOP point anymore.}

\item{LOLLIPOPMouseOutAnimationTime}{Animation time when mouse moving out the element}

\item{LOLLIPOPMouseOutColor}{Color when mouse moving out the element}

\item{LOLLIPOPMouseOutCircleSize}{Circle size when mouse moving out the element}

\item{LOLLIPOPMouseOutCircleOpacity}{Circle opacity when mouse moving out the element}

\item{LOLLIPOPMouseOutCircleStrokeColor}{Circle stroke color when mouse moving out the element}

\item{LOLLIPOPMouseOutCircleStrokeWidth}{Circle stroke width when mouse moving out the element}

\item{LOLLIPOPMouseUpDisplay}{Default False, show/not the tooltip when mouse click up a LOLLIPOP point.}

\item{LOLLIPOPMouseUpColor}{Color when mouse moving up the element}

\item{LOLLIPOPMouseUpCircleSize}{Circle size when mouse moving up the element}

\item{LOLLIPOPMouseUpCircleOpacity}{Circle opacity when mouse moving up the element}

\item{LOLLIPOPMouseUpCircleStrokeColor}{Circle stroke color when mouse moving up the element}

\item{LOLLIPOPMouseUpCircleStrokeWidth}{Circle stroke width when mouse moving up the element}

\item{LOLLIPOPMouseOverDisplay}{Default False, show/not the tooltip when mouse hover on a LOLLIPOP point.}

\item{LOLLIPOPMouseOverColor}{Color when mouse moving over the element}

\item{LOLLIPOPMouseOverCircleSize}{Circle size when mouse moving over the element}

\item{LOLLIPOPMouseOverCircleOpacity}{Circle opacity when mouse moving over the element}

\item{LOLLIPOPMouseOverCircleStrokeColor}{Circle stroke color when mouse moving over the element}

\item{LOLLIPOPMouseOverCircleStrokeWidth}{Circle stroke width when mouse moving over the element}

\item{LOLLIPOPMouseOverTooltipsSetting}{Default "style1"}

\item{LOLLIPOPMouseOverTooltipsHtml}{Default " "}

\item{LOLLIPOPMouseOverTooltipsPosition}{Default "absolute"}

\item{LOLLIPOPMouseOverTooltipsBackgroundColor}{Default "white"}

\item{LOLLIPOPMouseOverTooltipsBorderStyle}{Default "solid"}

\item{LOLLIPOPMouseOverTooltipsBorderWidth}{Default 0}

\item{LOLLIPOPMouseOverTooltipsPadding}{Default "3px"}

\item{LOLLIPOPMouseOverTooltipsBorderRadius}{Default "3px"}

\item{LOLLIPOPMouseOverTooltipsOpacity}{Default 0.8}

\item{elementId}{the name of the HTML id to be used to contain the visualization.}

\item{...}{Ignored}
}
\description{
A R packages based on Next Generation Circos
}
\examples{
NGCircos(genome = "hg19")

}
