% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCircos.R
\name{NGCircosArcTrack}
\alias{NGCircosArcTrack}
\title{Create a ARC module to a NGCircos tracklist}
\usage{
NGCircosArcTrack(
  trackname,
  compareGroup = 1,
  outerRadius = 150,
  innerRadius = 130,
  opacity = 1,
  animationDisplay = FALSE,
  animationTime = 2000,
  animationDelay = 20,
  animationType = "bounce",
  data,
  ...
)
}
\arguments{
\item{trackname}{The name of the new track.}

\item{compareGroup}{The group number of thic track in compare module}

\item{innerRadius, outerRadius}{Where the track should begin and end}

\item{opacity}{The opacity for arc}

\item{animationDisplay}{Whether display animation}

\item{animationTime, animationDelay, animationType}{The time, delay and display type for animation}

\item{data}{A list of arc with details including chr, start, end, color, des, link and html.
Details can be found on NGCircos document.}

\item{...}{Ignored}
}
\description{
Display the CNV without value, Gene domain, Chromosome band in the visualization
}
\examples{
arcData<-arcExample
NGCircos(NGCircosArcTrack('ArcTrack01', outerRadius = 212, innerRadius = 224, data=arcData),
 genome=list("EGFR"=1211),outerRadius = 220,genomeFillColor = c("grey"))

}
