% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCircos.R
\name{NGCircosBackgroundTrack}
\alias{NGCircosBackgroundTrack}
\title{Create a BACKGROUND track to be added to a NGCircos tracklist}
\usage{
NGCircosBackgroundTrack(
  trackname,
  compareGroup = 1,
  fillColors = "#EEEEFF",
  borderColors = "#000000",
  axisShow = FALSE,
  axisColor = "#000",
  axisOpacity = 0.5,
  axisNum = 4,
  axisWidth = 0.3,
  maxRadius = 190,
  minRadius = 105,
  borderSize = 0.3,
  animationDisplay = FALSE,
  animationTime = 2000,
  animationDelay = 20,
  animationType = "bounce",
  ...
)
}
\arguments{
\item{trackname}{The name of the new track.}

\item{compareGroup}{The group number of thic track in compare module}

\item{fillColors}{The color of the background element, in hexadecimal RGB format.}

\item{borderColors}{The color of the background borders, in hexadecimal RGB format.}

\item{axisShow}{Whether show a axis or not}

\item{axisWidth, axisColor, axisOpacity, axisNum}{The color, opacity value and number of line for axis}

\item{minRadius, maxRadius}{Where the track should begin and end}

\item{borderSize}{The thickness of the background borders.}

\item{animationDisplay}{Whether display a animation or not}

\item{animationTime, animationDelay, animationType}{The time, delay and display type for animation}

\item{...}{Ignored}
}
\description{
Simple background to display behind another track
}
\examples{
NGCircos(NGCircosBackgroundTrack('bgTrack01', fillColors="#FFEEEE", borderSize = 1))

}
