% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCircos.R
\name{NGCircosModuleList}
\alias{NGCircosModuleList}
\alias{+.NGCircosModuleList}
\alias{-.NGCircosModuleList}
\title{Create a list of NGCircos modules}
\usage{
NGCircosModuleList()

\method{+}{NGCircosModuleList}(x, ...)

\method{-}{NGCircosModuleList}(x, ...)
}
\arguments{
\item{x}{The moduleList on which other modules should be added or removed.}

\item{...}{The modules to add (as moduleLists) or to remove (as module names).}
}
\description{
This allows the use of the '+' and '-' operator on these lists
}
