% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGCircos.R
\name{NGCircosTracklist}
\alias{NGCircosTracklist}
\alias{+.NGCircosTracklist}
\alias{-.NGCircosTracklist}
\title{Create a list of NGCircos tracks}
\usage{
NGCircosTracklist()

\method{+}{NGCircosTracklist}(x, ...)

\method{-}{NGCircosTracklist}(x, ...)
}
\arguments{
\item{x}{The tracklist on which other tracks should be added or removed.}

\item{...}{The tracks to add (as tracklists) or to remove (as track names).}
}
\description{
This allows the use of the '+' and '-' operator on these lists
}
