% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNGCircos.R
\name{rNGCircosArc}
\alias{rNGCircosArc}
\title{Create a ARC module to a moduleList}
\usage{
rNGCircosArc(
  modulename,
  compareGroup = 1,
  outerRadius = 150,
  innerRadius = 130,
  opacity = 1,
  animationDisplay = FALSE,
  animationTime = 2000,
  animationDelay = 20,
  animationType = "bounce",
  data,
  ...
)
}
\arguments{
\item{modulename}{The name of the new module.}

\item{compareGroup}{The group number of this module in compare module}

\item{innerRadius, outerRadius}{Where the module should begin and end}

\item{opacity}{The opacity for arc}

\item{animationDisplay}{Whether display animation}

\item{animationTime, animationDelay, animationType}{The time, delay and display type for animation}

\item{data}{A list of arc with details including chr, start, end, color, des, link and html.
Details can be found on document.}

\item{...}{Ignored}
}
\description{
Display the CNV without value, Gene domain, Chromosome band in the visualization
}
\examples{
arcData<-arcExample
rNGCircos(rNGCircosArc('Arc01', outerRadius = 212, innerRadius = 224, data=arcData),
 genome=list("EGFR"=1211),outerRadius = 220,genomeFillColor = c("grey"))

}
