% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\docType{data}
\name{snpExample}
\alias{snpExample}
\title{Snp plot example data}
\format{
A data frame with 10 columns:
\describe{
  \item{chr}{chromosome}
  \item{pos}{position}
  \item{value}{value,such as p-value}
  \item{des}{description}
  \item{color}{color}
  \item{r2value}{r2 value}
  \item{link}{hyperlink for snp}
  \item{index}{index for combination}
  \item{image}{image for combination}
  \item{html}{The external html language}
}
}
\usage{
snpExample
}
\description{
The data is in matrix with column names
}
\keyword{datasets}
