% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interacCircos.R
\docType{data}
\name{wigExample}
\alias{wigExample}
\title{Wig plot example data}
\format{
A data frame with 5 columns:
\describe{
  \item{chr}{chromosome}
  \item{pos}{position}
  \item{des}{description}
  \item{value}{value}
  \item{html}{The external html language}
}
}
\usage{
wigExample
}
\description{
The data is in matrix with column names
}
\keyword{datasets}
