% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rNGCircos.R
\docType{data}
\name{chordExample}
\alias{chordExample}
\title{Chord plot example data}
\format{
A data frame in which each value represents the relationship from a column to a row:
\describe{
  \item{column name}{the name for each arc}
  \item{row}{the order and number is same as column, representing the same items}
}
}
\usage{
chordExample
}
\description{
The data is in matrix with row names
}
\keyword{datasets}
